### 2024-10-22 #7392 @kyleniemeyer
Hi @prashjha, can you address this missing DOI messages above?

@matthewfeickert could you edit this submission?

Just for full disclosure, @prashjha is a JOSS editor.
