+++ docs/joss/paper.bib
@@ -1,22 +1,26 @@
 @article{cundall1979discrete,
-  title={A discrete numerical model for granular assemblies},
-  author={Cundall, Peter A and Strack, Otto DL},
-  journal={geotechnique},
-  volume={29},
-  number={1},
-  pages={47--65},
-  year={1979},
-  publisher={Thomas Telford Ltd}
+  author = {Cundall, P. A. and Strack, O. D. L.},
+  title = {A discrete numerical model for granular assemblies},
+  journal = {Géotechnique},
+  volume = {29},
+  number = {1},
+  pages = {47-65},
+  year = {1979},
+  doi = {10.1680/geot.1979.29.1.47},
+  url = {https://doi.org/10.1680/geot.1979.29.1.47}
 }
 
 @article{jha2021peridynamics,
-  title={Peridynamics-based discrete element method (PeriDEM) model of granular systems involving breakage of arbitrarily shaped particles},
-  author={Jha, Prashant K and Desai, Prathamesh S and Bhattacharya, Debdeep and Lipton, Robert},
-  journal={Journal of the Mechanics and Physics of Solids},
-  volume={151},
-  pages={104376},
-  year={2021},
-  publisher={Elsevier}
+  title = {Peridynamics-based discrete element method (PeriDEM) model of granular systems involving breakage of arbitrarily shaped particles},
+  journal = {Journal of the Mechanics and Physics of Solids},
+  volume = {151},
+  pages = {104376},
+  year = {2021},
+  issn = {0022-5096},
+  doi = {https://doi.org/10.1016/j.jmps.2021.104376},
+  url = {https://www.sciencedirect.com/science/article/pii/S0022509621000661},
+  author = {Prashant K. Jha and Prathamesh S. Desai and Debdeep Bhattacharya and Robert Lipton},
+  keywords = {Peridynamics, Discrete element method, Particle attrition, Particle interlocking, Particle breakage, Granular media, Fracture}
 }
 
 @incollection{silling2016introduction,
@@ -29,25 +33,28 @@
 }
 
 @article{silling2000reformulation,
-  title={Reformulation of elasticity theory for discontinuities and long-range forces},
-  author={Silling, Stewart A},
-  journal={Journal of the Mechanics and Physics of Solids},
-  volume={48},
-  number={1},
-  pages={175--209},
-  year={2000},
-  publisher={Elsevier}
+  title = {Reformulation of elasticity theory for discontinuities and long-range forces},
+  journal = {Journal of the Mechanics and Physics of Solids},
+  volume = {48},
+  number = {1},
+  pages = {175-209},
+  year = {2000},
+  issn = {0022-5096},
+  doi = {https://doi.org/10.1016/S0022-5096(99)00029-0},
+  url = {https://www.sciencedirect.com/science/article/pii/S0022509699000290},
+  author = {S.A. Silling}
 }
 
 @article{jha2018numerical,
- author = {Jha, Prashant K and Lipton, Robert},
- journal = {SIAM Journal on Numerical Analysis},
- number = {2},
- pages = {906--941},
- publisher = {SIAM},
- title = {Numerical analysis of nonlocal fracture models in Holder space},
- volume = {56},
- year = {2018}
+  author = {Jha, Prashant K and Lipton, Robert},
+  journal = {SIAM Journal on Numerical Analysis},
+  number = {2},
+  pages = {906--941},
+  publisher = {SIAM},
+  title = {Numerical analysis of nonlocal fracture models in Holder space},
+  volume = {56},
+  year = {2018,
[…]
