+++ docs/joss/paper.bib
@@ -53,7 +53,7 @@
   publisher = {SIAM},
   title = {Numerical analysis of nonlocal fracture models in Holder space},
   volume = {56},
-  year = {2018,
+  year = {2018},
   doi = {https://doi.org/10.1137/17M1112236}
 }
 
