+++ docs/joss/paper.bib
@@ -17,8 +17,8 @@
   pages = {104376},
   year = {2021},
   issn = {0022-5096},
-  doi = {https://doi.org/10.1016/j.jmps.2021.104376},
-  url = {https://www.sciencedirect.com/science/article/pii/S0022509621000661},
+  doi = {10.1016/j.jmps.2021.104376},
+  url = {https://doi.org/10.1016/j.jmps.2021.104376},
   author = {Prashant K. Jha and Prathamesh S. Desai and Debdeep Bhattacharya and Robert Lipton},
   keywords = {Peridynamics, Discrete element method, Particle attrition, Particle interlocking, Particle breakage, Granular media, Fracture}
 }
@@ -40,8 +40,8 @@
   pages = {175-209},
   year = {2000},
   issn = {0022-5096},
-  doi = {https://doi.org/10.1016/S0022-5096(99)00029-0},
-  url = {https://www.sciencedirect.com/science/article/pii/S0022509699000290},
+  doi = {10.1016/S0022-5096(99)00029-0},
+  url = {https://doi.org/10.1016/S0022-5096(99)00029-0},
   author = {S.A. Silling}
 }
 
@@ -54,7 +54,8 @@
   title = {Numerical analysis of nonlocal fracture models in Holder space},
   volume = {56},
   year = {2018},
-  doi = {https://doi.org/10.1137/17M1112236}
+  doi = {10.1137/17M1112236},
+url = {https://doi.org/10.1137/17M1112236}
 }
 
 @article{jha2019numerical,
@@ -65,7 +66,8 @@
  title = {Numerical convergence of finite difference approximations for state based peridynamic fracture models},
  volume = {351},
  year = {2019},
-doi = {https://doi.org/10.1016/j.cma.2019.03.024}
+doi = {10.1016/j.cma.2019.03.024},
+url = {https://doi.org/10.1016/j.cma.2019.03.024}
 }
 
 @article{jha2018numerical2,
@@ -77,7 +79,8 @@ doi = {https://doi.org/10.1016/j.cma.2019.03.024}
  title = {Numerical convergence of nonlinear nonlocal continuum models to local elastodynamics},
  volume = {114},
  year = {2018},
-doi = {https://doi.org/10.1002/nme.5791}
+doi = {10.1002/nme.5791},
+url = {https://doi.org/10.1002/nme.5791}
 }
 
 @article{Jha2020peri,
@@ -99,7 +102,8 @@ doi = {https://doi.org/10.1002/nme.5791}
  title = {Complex fracture nucleation and evolution with nonlocal elastodynamics},
  volume = {1},
  year = {2019},
-doi = {https://doi.org/10.1007/s42102-019-00010-0}
+doi = {10.1007/s42102-019-00010-0},
+url = {https://doi.org/10.1007/s42102-019-00010-0}
 }
 
 @article{jha2021finite,
@@ -111,7 +115,8 @@ doi = {https://doi.org/10.1007/s42102-019-00010-0}
   pages={1675},
   year={2021},
   publisher={American Institute of Mathematical Sciences},
-doi = {https://doi.org/10.3934/dcdsb.2020178}
+doi = {10.3934/dcdsb.2020178},
+url = {https://doi.org/10.3934/dcdsb.2020178}
 }
 
 @article{lipton2014dynamic,
@@ -122,7 +127,8 @@ doi = {https://doi.org/10.3934/dcdsb.2020178}
   pages={21--50},
   year={2014},
   publisher={Springer},
-doi = {https://doi.org/10.1007/s10659-013-9463-0}
[…]
