### 2024-10-28 4474a8d
Update paper.bib
