+++ docs/joss/paper.bib
@@ -19,13 +19,13 @@
   issn = {0022-5096},
   doi = {10.1016/j.jmps.2021.104376},
   url = {https://doi.org/10.1016/j.jmps.2021.104376},
-  author = {Prashant K. Jha and Prathamesh S. Desai and Debdeep Bhattacharya and Robert Lipton},
+  author = {Jha, Prashant K. and Desai, Prathamesh S. and Bhattacharya, Debdeep and Lipton, Robert},
   keywords = {Peridynamics, Discrete element method, Particle attrition, Particle interlocking, Particle breakage, Granular media, Fracture}
 }
 
 @incollection{silling2016introduction,
   title={Introduction to peridynamics},
-  author={Silling, Stewart A},
+  author={Silling, Stewart A.},
   booktitle={Handbook of peridynamic modeling},
   pages={63--98},
   year={2016},
@@ -42,11 +42,11 @@
   issn = {0022-5096},
   doi = {10.1016/S0022-5096(99)00029-0},
   url = {https://doi.org/10.1016/S0022-5096(99)00029-0},
-  author = {S.A. Silling}
+  author = {Silling, S. A.}
 }
 
 @article{jha2018numerical,
-  author = {Jha, Prashant K and Lipton, Robert},
+  author = {Jha, Prashant K. and Lipton, Robert},
   journal = {SIAM Journal on Numerical Analysis},
   number = {2},
   pages = {906--941},
@@ -59,7 +59,7 @@ url = {https://doi.org/10.1137/17M1112236}
 }
 
 @article{jha2019numerical,
- author = {Jha, Prashant K and Lipton, Robert},
+ author = {Jha, Prashant K. and Lipton, Robert},
  journal = {Computer Methods in Applied Mechanics and Engineering},
  pages = {184--225},
  publisher = {Elsevier},
@@ -71,7 +71,7 @@ url = {https://doi.org/10.1016/j.cma.2019.03.024}
 }
 
 @article{jha2018numerical2,
- author = {Jha, Prashant K and Lipton, Robert},
+ author = {Jha, Prashant K. and Lipton, Robert},
  journal = {International Journal for Numerical Methods in Engineering},
  number = {13},
  pages = {1389--1410},
@@ -84,7 +84,7 @@ url = {https://doi.org/10.1002/nme.5791}
 }
 
 @article{Jha2020peri,
- author = {Jha, Prashant K and Lipton, Robert P},
+ author = {Jha, Prashant K. and Lipton, Robert P},
  journal = {International Journal of Fracture},
  title = {Kinetic relations and local energy balance for LEFM from a nonlocal peridynamic model},
  year = {2020},
@@ -94,7 +94,7 @@ url = {https://doi.org/10.1002/nme.5791}
 }
 
 @article{lipton2019complex,
- author = {Lipton, Robert P and Lehoucq, Richard B and Jha, Prashant K},
+ author = {Lipton, Robert P. and Lehoucq, Richard B. and Jha, Prashant K.},
  journal = {Journal of Peridynamics and Nonlocal Modeling},
  number = {2},
  pages = {122--130},
@@ -108,7 +108,7 @@ url = {https://doi.org/10.1007/s42102-019-00010-0}
 
 @article{jha2021finite,
   title={Finite element approximation of nonlocal dynamic fracture models},
-  author={Jha, PK and Lipton, R},
+  author={Jha, Prashant K. and Lipton, R.},
   journal={Discrete \& Continuous Dynamical Systems-B},
   volume={26},
   number={3},
@@ -140,7 +140,7 @@ url = {https://hdl.handle.net/11299/215346}
 
 @article{silling2007peridynamic,
   title={Peridynamic states and constitutive modeling},
[…]
