### 2024-11-25 #7525 @editorialbot
[REVIEW]: PeriDEM - High-fidelity modeling of granular media consisting of deformable complex-shaped particles
**Submitting author:** @prashjha (<a href="http://orcid.org/0000-0003-2158-364X">Prashant K Jha</a>)
**Repository:** https://github.com/prashjha/PeriDEM
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.2
