### 2024-11-25 #7525 @matthewfeickert
As mentioned in https://github.com/openjournals/joss-reviews/issues/7392#issuecomment-2497367132, we have had some additional potential reviewers respond to agree to also contribute reviews, and I like to accept 3 reviewers when possible. So @ziyixi will be joining as a reviewer (thanks!).

@ziyixi, please check out https://github.com/openjournals/joss-reviews/issues/7525#issuecomment-2497387875 for additional information, and if you have any questions please let me know.
