### 2024-12-09 prashjha/PeriDEM#35 @nrichart
Failing Test: Two-particle with wall
## Expected Behavior

Running test `Two-particle with wall`

## Actual Behavior

```
Error: For this particle/wall in zone = Zone_3 , neither config['Particle'] nor config['Wall'] exist in input file.
```

## Steps to Reproduce the Problem

I followed the step from the documentation

## Specifications

  - HPX Version: not part of the dependencies
  - Platform (compiler, OS): debian testing
