### 2024-12-09 prashjha/PeriDEM#36 @nrichart
Failling Test: Compression test
## Expected Behavior

running the test `Compressive test`

## Actual Behavior

```
terminate called after throwing an instance of 'YAML::BadFile'
  what():  bad file: input_0.yaml

PeriDEM:190808 terminated with signal 6 at PC=7f78ea49de5c SP=7ffc45d99ea0.  Backtrace:
/lib/x86_64-linux-gnu/libc.so.6(+0x93e5c) [0x7f78ea49de5c]
/lib/x86_64-linux-gnu/libc.so.6(gsignal+0x12) [0x7f78ea449c82]
/lib/x86_64-linux-gnu/libc.so.6(abort+0xd3) [0x7f78ea4324f0]
/lib/x86_64-linux-gnu/libstdc++.so.6(+0xa1a3d) [0x7f78ea6a1a3d]
/lib/x86_64-linux-gnu/libstdc++.so.6(+0xb300a) [0x7f78ea6b300a]
/lib/x86_64-linux-gnu/libstdc++.so.6(_ZSt10unexpectedv+0x0) [0x7f78ea6a15e9]
/lib/x86_64-linux-gnu/libstdc++.so.6(+0xb3288) [0x7f78ea6b3288]
/home/richart/dev/perso/PeriDEM/build/bin/PeriDEM(+0x17f17) [0x55c097d0af17]
/home/richart/dev/perso/PeriDEM/build/bin/PeriDEM(+0x9395e) [0x55c097d8695e]
/home/richart/dev/perso/PeriDEM/build/bin/PeriDEM(+0xa7527) [0x55c097d9a527]
/home/richart/dev/perso/PeriDEM/build/bin/PeriDEM(+0x259aa) [0x55c097d189aa]
/lib/x86_64-linux-gnu/libc.so.6(+0x29d68) [0x7f78ea433d68]
/lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0x85) [0x7f78ea433e25]
/home/richart/dev/perso/PeriDEM/build/bin/PeriDEM(+0x27ce1) [0x55c097d1ace1]
```

## Steps to Reproduce the Problem

the steps from the documentation

## Specifications

  - Platform (compiler, OS): debian testing
