### 2024-12-09 prashjha/PeriDEM#37 @nrichart
My points for the JOSS review
Hello,

I am creating an issue to list the thing I noticed while  testing the code for the JOSS review https://github.com/openjournals/joss-reviews/issues/7525

I will list the in order of importance from my point of view from most to least important

- The tests documented in the main README do not work for me I opened 2 specific issues #35 #36
- The documentation does not have a `Contributing` section
- There is no `installation` part only a `build` instruction, the code has to be run from the build folder
- The generated documentation does not show properly https://prashjha.github.io/PeriDEM/
  -  `<a href=` and `@ref`
  - the link for the examples do not work they point to github.io instead of to the source
  - in tools/README.md you mention dependencies that are not in the main README.md like `libfann` you also install more than the list of dependencies in the brew and apt commands like for example `tbb`
- The build instruction cannot be copy/pasted due to the comment in the `cmake` command
- There is no mention that the `cmake` and `make` command should/could be run in a build folder

--
Nicolas
