### 2024-12-09 prashjha/PeriDEM#37 @prashjha
Hi @nrichart,

Thank you for the comments.

The failing tests are due to significant changes. I will rectify them and let you know.
- On `installation`, I will see if I can modify cmake to include the install option.
- I found it hard to work with links in doxygen. I wanted to display the same files in README.md and doxygen; I could not do this properly with doxygen. I will see if I can correct these errors as well.
- I will address other comments also in due time.

Best,
Prashant
