### 2024-12-10 prashjha/PeriDEM#37 @nrichart
@prashjha, thanks.
For the doc the main point I think are the section titles that are also links, since this seams to generate 2 errors in the doxygen. It does not show properly but it also mess up the table of content.
The other links I understand your problem, to not have to maintain 2 places with the same content. Could full link help instead of relative ones ?
Best,
--
Nicolas
