### 2025-01-07 #7525 @matthewfeickert
@divijghose, @nrichart, @ziyixi Happy 2025! I wanted to check in to see how the review is going. If you can all give some short updates that would be useful, though as it is the start of the year I'll also set @editorialbot reminders for us all.

Thanks very much for opening up Issues for the review @nrichart. These are great and I hope that the other reviewers will open similar ones if needed. From looking at https://github.com/prashjha/PeriDEM/issues/35 and https://github.com/prashjha/PeriDEM/issues/36, I assume that @nrichart's update is waiting for author fixes (which is reasonable as the end of the year and holidays slows everything down), but if there's additional info please share also. :+1:
