### 2025-01-14 #7525 @nrichart
@matthewfeickert, happy 2025 to you too. Yes I was waiting for some updates on the PRs, I have a few remarks on the paper too. I will try to take some time this week to refresh my notes and add a comment.
