### 2025-02-06 ad056cf
JOSS review: clean-up, update example files from test/test_data/peridem
