+++ examples/PeriDEM/two_particles/twop_circ/inp/problem_setup.py
@@ -70,22 +70,8 @@ def particle_locations(inp_dir, pp_tag, R1, R2, offset):
   """Generate particle location data"""
 
   sim_particles = []
-  sim_particles.append([0., R1, R1, 0., R1])
-  sim_particles.append([1., R1, 2. * R1 + R2 + offset, 0., R2])
-
-  inpf = open(inp_dir + 'particle_locations_' + str(pp_tag) + '.csv','w')
-  inpf.write("i, x, y, z, r\n")
-  for p in sim_particles:
-    inpf.write("%d, %Lf, %Lf, %Lf, %Lf\n" % (int(p[0]), p[1], p[2], p[3], p[4]))
-
-  inpf.close()
-
-def particle_locations_orient(inp_dir, pp_tag, R1, R2, offset):
-  """Generate particle location data"""
-
-  sim_particles = []
-  sim_particles.append([0., R1, R1, 0., R1, 0.5*np.pi])
-  sim_particles.append([1., R1, 2. * R1 + R2 + offset, 0., R2, 0.])
+  sim_particles.append([0., R1, R1, 0., R1, 0.])
+  sim_particles.append([1., R1, 2. * R1 + R2 + offset, 0., R2, np.pi*0.5])
 
   inpf = open(inp_dir + 'particle_locations_' + str(pp_tag) + '.csv','w')
   inpf.write("i, x, y, z, r, o\n")
@@ -94,117 +80,19 @@ def particle_locations_orient(inp_dir, pp_tag, R1, R2, offset):
 
   inpf.close()
 
-def rotate(p, theta, axis):
 
-  p_np = np.array(p)
-  axis_np = np.array(axis)
-
-  ct = np.cos(theta);
-  st = np.sin(theta);
-
-  # dot
-  p_dot_n = np.dot(p_np,axis_np)
-
-  # cross
-  n_cross_p = np.cross(axis_np, p_np)
-
-  return (1. - ct) * p_dot_n * axis_np + ct * p_np + st * n_cross_p
-
-
-def get_ref_hex_points(center, radius, add_center = False):
-
-  # drum2d
-  #
-  #                        v3           v2
-  #                         +           +
-  #
-  #
-  #                      +         o           +
-  #                     v4         x            v1
-  #
-  #                         +           +
-  #                        v5           v6
-  #
-  # Axis is a vector from x to v1
-  #
-
-  axis = [1., 0., 0.]
-
-  # center and radius
-  sim_Cx = center[0]
-  sim_Cy = center[1]
-  sim_Cz = center[2]
-  sim_radius = radius
-
-  rotate_axis = [0., 0., 1.]
-
-  points = []
-  if add_center:
-    points.append(center)
-
-  for i in range(6):
-  	xi = rotate(axis, i*np.pi/3., rotate_axis)
-  	points.append([center[i] + radius * xi[i] for i in range(3)])
-
-  return points
-
-
-def get_ref_drum_points(center, radius, width, add_center = False):
-
-  # drum2d
-  #
-  #             v3                                v2
-  #               +                               +
-  #
-  #
-  #                      +         o           +
-  #                     v4         x            v1
-  #
-  #               +                                +
-  #               v5                               v6
-  #
-  # Axis is a vector from x to v1
-  #
-
[…]
