+++ examples/PeriDEM/two_particles/twop_wall_concave_diff_material_diff_size/inp/problem_setup.py
@@ -65,13 +65,18 @@ def get_G(E, nu):
 def get_eff_k(k1, k2):
   return 2. * k1 * k2 / (k1 + k2)
 
+def rect_to_five_param(r):
+  return [r[3] - r[0], r[4] - r[1], 0.5*(r[0] + r[3]), 0.5*(r[1] + r[4]), 0.5*(r[2] + r[5])]
 
-def particle_locations(inp_dir, pp_tag, R1, R2, offset):
+def particle_locations(inp_dir, pp_tag, R1, R2, offset, w1_rect_five_format, w2_rect_five_format):
   """Generate particle location data"""
 
   sim_particles = []
   sim_particles.append([0., R1, R1, 0., R1])
   sim_particles.append([1., R1, 2. * R1 + R2 + offset, 0., R2])
+
+  sim_particles.append([2., w1_rect_five_format[2], w1_rect_five_format[3], w1_rect_five_format[4], w1_rect_five_format[0]])
+  sim_particles.append([3., w2_rect_five_format[2], w2_rect_five_format[3], w2_rect_five_format[4], w2_rect_five_format[0]])
 
   inpf = open(inp_dir + 'particle_locations_' + str(pp_tag) + '.csv','w')
   inpf.write("i, x, y, z, r\n")
@@ -80,12 +85,15 @@ def particle_locations(inp_dir, pp_tag, R1, R2, offset):
 
   inpf.close()
 
-def particle_locations_orient(inp_dir, pp_tag, R1, R2, offset):
+def particle_locations_orient(inp_dir, pp_tag, R1, R2, offset, w1_rect_five_format, w2_rect_five_format):
   """Generate particle location data"""
 
   sim_particles = []
   sim_particles.append([0., R1, R1, 0., R1, 0.5*np.pi])
-  sim_particles.append([1., R1, 2. * R1 + R2 + offset, 0., R2, 0.])
+  sim_particles.append([1., R1+1.*R1*np.sin(np.pi/3.), 2. * R1 + R2 + offset, 0., R2, 0.3*np.pi])
+
+  sim_particles.append([2., w1_rect_five_format[2], w1_rect_five_format[3], w1_rect_five_format[4], w1_rect_five_format[0], 0.])
+  sim_particles.append([3., w2_rect_five_format[2], w2_rect_five_format[3], w2_rect_five_format[4], w2_rect_five_format[0], 0.])
 
   inpf = open(inp_dir + 'particle_locations_' + str(pp_tag) + '.csv','w')
   inpf.write("i, x, y, z, r, o\n")
@@ -109,44 +117,6 @@ def rotate(p, theta, axis):
   n_cross_p = np.cross(axis_np, p_np)
 
   return (1. - ct) * p_dot_n * axis_np + ct * p_np + st * n_cross_p
-
-
-def get_ref_hex_points(center, radius, add_center = False):
-
-  # drum2d
-  #
-  #                        v3           v2
-  #                         +           +
-  #
-  #
-  #                      +         o           +
-  #                     v4         x            v1
-  #
-  #                         +           +
-  #                        v5           v6
-  #
-  # Axis is a vector from x to v1
-  #
-
-  axis = [1., 0., 0.]
-
-  # center and radius
-  sim_Cx = center[0]
-  sim_Cy = center[1]
-  sim_Cz = center[2]
[…]

[…]
