+++ src/geometry/CMakeLists.txt
@@ -10,4 +10,4 @@ AUX_SOURCE_DIRECTORY(./ SOURCES)
 
 add_library(Geometry ${SOURCES})
 
-target_link_libraries(Geometry PUBLIC NSearch fmt::fmt Threads::Threads)
+target_link_libraries(Geometry PUBLIC Util NSearch fmt::fmt Threads::Threads)
