+++ search/pages_8.js
@@ -1,7 +1,0 @@
-var searchData=
-[
-  ['parallelcomp_0',['parallelcomp',['../md__2home_2user_2project_2test_2test__data_2parallelcomp_2README.html',1,'']]],
-  ['particles_1',['PeriDEM - High-fidelity modeling of granular media consisting of deformable complex-shaped particles',['../index.html',1,'']]],
-  ['peridem_20high_20fidelity_20modeling_20of_20granular_20media_20consisting_20of_20deformable_20complex_20shaped_20particles_2',['PeriDEM - High-fidelity modeling of granular media consisting of deformable complex-shaped particles',['../index.html',1,'']]],
-  ['peridem_3a_20installation_3',['PeriDEM: Installation',['../md__2home_2user_2project_2tools_2README.html',1,'']]]
-];
