+++ dir_dd8088c393ec7a5708348b4337ae0957.js
@@ -1,4 +1,0 @@
-var dir_dd8088c393ec7a5708348b4337ae0957 =
-[
-    [ "problem_setup.py", "test_2test__data_2peridem_2twop__circ__diff__radius_2inp_2problem__setup_8py.html", "test_2test__data_2peridem_2twop__circ__diff__radius_2inp_2problem__setup_8py" ]
-];
