+++ search/all_18.js
@@ -1,6 +1,0 @@
-var searchData=
-[
-  ['_7elogger_0',['~Logger',['../classutil_1_1io_1_1Logger.html#af442dc23f6ae5f0617b5653c8a8d9055',1,'util::io::Logger']]],
-  ['_7ematerial_1',['~Material',['../classmaterial_1_1Material.html#a615c87e536d89cd54570bb535249469a',1,'material::Material']]],
-  ['_7ewriter_2',['~Writer',['../classrw_1_1writer_1_1Writer.html#a608705abe82070580659af4bbb0c6b68',1,'rw::writer::Writer']]]
-];
