+++ search/classes_11.js
@@ -1,7 +1,0 @@
-var searchData=
-[
-  ['vtkparticlereader_0',['VtkParticleReader',['../classrw_1_1reader_1_1VtkParticleReader.html',1,'rw::reader']]],
-  ['vtkparticlewriter_1',['VtkParticleWriter',['../classrw_1_1writer_1_1VtkParticleWriter.html',1,'rw::writer']]],
-  ['vtkreader_2',['VtkReader',['../classrw_1_1reader_1_1VtkReader.html',1,'rw::reader']]],
-  ['vtkwriter_3',['VtkWriter',['../classrw_1_1writer_1_1VtkWriter.html',1,'rw::writer']]]
-];

[…]
