### 2025-02-07 6d94357
JOSS review: finalize examples
