+++ README.md
@@ -47,7 +47,7 @@ We have created channels on various platforms:
 - [PeriDEM on slack](peridem.slack.com)
   * Send us an email if interested in joining the workspace.
 
-## [Documentation](https://prashjha.github.io/PeriDEM/)
+## Documentation
 [Doxygen generated documentation](https://prashjha.github.io/PeriDEM/) details functions and objects in the library. 
 
 ## Tutorial
@@ -55,24 +55,24 @@ We explain the setting-up of simulations in further details in [tutorial](./tuto
 We consider `two-particle` test setup with non-circular particles and `compressive-test` to 
 discuss the various aspects of simulations.
 
-## [Examples](./examples/README.md)
+## Examples
 We next highlight some key examples. Further details are available in [examples/README.md](./examples/README.md). 
 
 ### Two-particle tests
 
-|       <img src="./assets/two_particle_circ_no_damp.gif" width="200">       |     <img src="./assets/two_particle_circ_damp.gif" width="200">      |
-|:--------------------------------------------------------------------------:|:--------------------------------------------------------------------:|
-| [Circular without damping](./examples/PeriDEM/two_particles/circ_no_damp/) | [Circular with damping](./examples/PeriDEM/two_particles/circ_damp/) |
+|      <img src="./assets/two_particle_circ_no_damp.gif" width="200">       |     <img src="./assets/two_particle_circ_damp.gif" width="200">      |
+|:-------------------------------------------------------------------------:|:--------------------------------------------------------------------:|
+| Circular without damping | Circular with damping |
 
-|    <img src="./assets/two_particle_circ_diff_material.gif" width="200">     |   <img src="./assets/two_particle_circ_damp_diff_radius.gif" width="200">   |            <img src="./assets/two_particle_circ_diff_radius_diff_material.gif" width="200">             |
-|:---------------------------------------------------------------------------:|:---------------------------------------------------------------------------:|:-------------------------------------------------------------------------------------------------------:|
-| [Different materials](./examples/PeriDEM/two_particles/circ_diff_material/) | [Different radius](./examples/PeriDEM/two_particles/circ_damp_diff_radius/) | [Different radius different material](./examples/PeriDEM/two_particles/circ_diff_radius_diff_material/) |
+|    <img src="./assets/two_particle_circ_diff_material.gif" width="200">    |   <img src="./assets/two_particle_circ_damp_diff_radius.gif" width="200">   |            <img src="./assets/two_particle_circ_diff_radius_diff_material.gif" width="200">            |
+|:--------------------------------------------------------------------------:|:---------------------------------------------------------------------------:|:------------------------------------------------------------------------------------------------------:|
+| Different materials | Different radius | Different radius different material |
 
 ### Two-particle with wall test
 
-|   <img src="./assets/two_particle_wall_concave_diff_material_diff_size.gif" width="400">    | 
-|:-------------------------------------------------------------------------------------------:| 
-| [Concave particles](./examples/PeriDEM/two_particles_wall/concave_diff_material_diff_size/) |
+|   <img src="./assets/two_particle_wall_concave_diff_material_diff_size.gif" width="400">   | 
+|:------------------------------------------------------------------------------------------:| 
+| Concave particles |
 
 ### Compressive tests
 Setup for this test consists of 502 circular and hexagonal-shaped particles of varying 
[…]
