### 2025-02-07 prashjha/PeriDEM#37 @prashjha
Hello @nrichart,

In the branch 'fix-joss', I have addressed the following:

- Failing two particle test. I replaced the test with the new version in the directory `examples/PeriDEM/two_particles/twop_wall_concave_diff_material_diff_size`. I tested this and other examples in the example directory, and they all run. Here is the [output.log](https://github.com/user-attachments/files/18699753/output.log) for the `examples/PeriDEM/two_particles/twop_wall_concave_diff_material_diff_size` example.
- Failing compressive test. The new version is in the directory `examples/PeriDEM/compressive_test/compression_large_set`. Corresponding [output.log](https://github.com/user-attachments/files/18699780/output.log) is attached.
- I have modified README.md file to not have links in headlines. I hope this will fix some of the issues in generated documentation.

I will keep working on other issue and the issue of proper documentation in coming days.

Best,
Prashant
