+++ src/fe/mesh.cpp
@@ -1,422 +1,0 @@
-/*
- * -------------------------------------------
- * Copyright (c) 2021 - 2024 Prashant K. Jha
- * -------------------------------------------
- * PeriDEM https://github.com/prashjha/PeriDEM
- *
- * Distributed under the Boost Software License, Version 1.0. (See accompanying
- * file LICENSE)
- */
-
-#include "mesh.h"
-#include "inp/decks/meshDeck.h"
-#include "quadElem.h"
-#include "rw/reader.h"
-#include "tetElem.h"
-#include "triElem.h"
-#include "util/feElementDefs.h"
-#include "util/function.h"
-#include "util/geom.h"
-#include "util/parallelUtil.h"
-#include <cstdint>
-#include <iostream>
-#include <taskflow/taskflow/taskflow.hpp>
-#include <taskflow/taskflow/algorithm/for_each.hpp>
-
-fe::Mesh::Mesh(size_t dim)
-    : d_numNodes(0), d_numElems(0), d_eType(1), d_eNumVertex(0), d_numDofs(0),
-      d_h(0.), d_dim(dim), d_encDataPopulated(false), d_needEncData(false),
-      d_nPart(0){}
-
-fe::Mesh::Mesh(inp::MeshDeck *deck)
-    : d_numNodes(0), d_numElems(0), d_eType(1), d_eNumVertex(0), d_numDofs(0),
-      d_h(deck->d_h), d_dim(deck->d_dim),
-      d_spatialDiscretization(deck->d_spatialDiscretization),
-      d_filename(deck->d_filename), d_encDataPopulated(false),
-      d_needEncData(deck->d_populateElementNodeConnectivity),
-      d_nPart(0) {
-
-  // perform check on input data
-  if (d_spatialDiscretization != "finite_difference" and
-      d_spatialDiscretization != "weak_finite_element" and
-      d_spatialDiscretization != "nodal_finite_element" and
-      d_spatialDiscretization != "truss_finite_element") {
-    std::cerr << "Error: Spatial discretization type not known. Check input "
-                 "data.\n";
-    std::cerr << deck->printStr() << "\n";
-    exit(1);
-  }
-
-  if (d_dim < 0 or d_dim > 3) {
-    std::cerr << "Error: Check Dimension in input data.\n";
-    exit(1);
-  }
-
-  if (d_filename.empty()) {
-    std::cerr << "Error: Filename for mesh data not specified.\n";
-    exit(1);
-  }
-
-  // read mesh data from file
-  createData(d_filename);
-}
-
-//
-// Utility functions
-//
-void fe::Mesh::createData(const std::string &filename, bool ref_config) {
-
-  util::io::log("Mesh: Reading element data.\n");
-
-  int file_type = -1;
-  // find the extension of file and call correct reader
-  if (util::io::getExtensionFromFile(filename) == "csv")
-    file_type = 0;
-  else if (util::io::getExtensionFromFile(filename) == "msh")
-    file_type = 1;
-  else if (util::io::getExtensionFromFile(filename) == "vtu")
-    file_type = 2;
-  else {
-    std::cerr << "Error: Currently only '.csv', '.msg', and '.vtu' "
-                 "files are supported for reading mesh.\n";
-    exit(EXIT_FAILURE);
-  }
-
-  if (d_spatialDiscretization != "finite_difference" and file_type == 0) {
-
-    std::cerr << "Error: For discretization = " << d_spatialDiscretization
-              << " .vtu or .msh mesh file is required.\n";
-    exit(1);
-  }
-
-  //
-  bool is_fd = false;
-  if (d_spatialDiscretization == "finite_difference")
-    is_fd = true;
-
-  // read node and elements
-  if (file_type == 0)
[…]
