+++ search/files_3.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['elemincludes_2eh_0',['elemIncludes.h',['../elemIncludes_8h.html',1,'']]]
-];
