+++ search/pages_a.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['readme_0',['README',['../md__2home_2user_2project_2examples_2README.html',1,'']]]
+];
