### 2025-02-07 cca8621
JOSS review: add contributing section
