### 2025-02-07 a4b56da
JOSS review: fix paths in README.md files
