### 2025-02-07 5de9771
JOSS review: modify README.md so that cmake command can be copied
