### 2025-02-07 184ebbc
JOSS review: update README.md and make a remark on cmake, make, and installation
