### 2025-02-07 054af48
JOSS review: README.md update to mention additional libraries needed by dependencies
