+++ pNeighborDeck_8h.js
@@ -1,4 +1,0 @@
-var pNeighborDeck_8h =
-[
-    [ "inp::PNeighborDeck", "structinp_1_1PNeighborDeck.html", "structinp_1_1PNeighborDeck" ]
-];

[…]
