### 2025-02-07 prashjha/PeriDEM#37 @prashjha
Hi @nrichart, I think I have addressed all of your comments with the recent commits. We have some limitations:

1. In the generated documentation, README files in subdirectories mess up the paths to various contents. I will leave it as it is. Our generated documentation allows users to look up the code contents, e.g., functions and classes. In that sense, it serves the purpose.

Alternatively, I may skip rendering README files within subdirectories in the doxygen. Let me know your thoughts.

2. The install currently only installs the executable (I tested it, and it works). It will also be nice to be able to install headers and libraries in the future.

Let me know if this is satisfactory!
