+++ search/searchdata.js
@@ -1,45 +1,0 @@
-var indexSectionsWithContent =
-{
-  0: "abcdefghiklmnopqrstuvwyz~",
-  1: "abcdfghilmnopqrstvwz",
-  2: "adfgilmnprtu",
-  3: "bcdefgilmnopqrtuvwz",
-  4: "abcdefghiklmnopqrstuvwz~",
-  5: "abcdefghiklmnoprtuvw",
-  6: "cdilnprtu",
-  7: "o",
-  8: "c",
-  9: "egi",
-  10: "cdefghimoprst"
-};
-
-var indexSectionNames =
-{
-  0: "all",
-  1: "classes",
-  2: "namespaces",
-  3: "files",
-  4: "functions",
-  5: "variables",
-  6: "typedefs",
-  7: "related",
-  8: "defines",
-  9: "groups",
-  10: "pages"
-};
-
-var indexSectionLabels =
-{
-  0: "All",
-  1: "Data Structures",
-  2: "Namespaces",
-  3: "Files",
-  4: "Functions",
-  5: "Variables",
-  6: "Typedefs",
-  7: "Friends",
-  8: "Macros",
-  9: "Modules",
-  10: "Pages"
-};
-
