+++ search/files_1.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['contactdeck_2eh_0',['contactDeck.h',['../contactDeck_8h.html',1,'']]]
-];
