### 2025-02-10 prashjha/PeriDEM#37 @nrichart
Hi @prashjha,
Thank you for your work.
I will do my best to try this week to look at the code again and work on the review of your paper.
