### 2025-02-11 fb57f1d
consistent scaling for cuboid reference particle and particle
