+++ src/model/dem/demModel.cpp
@@ -1321,7 +1321,7 @@ void model::DEMModel::createParticlesFromFile(
 
     // create geometrical object
     std::shared_ptr<util::geometry::GeomObject> p_geom;
-    createGeometryAtSite(particle_radius,
+    auto scale = createGeometryAtSite(particle_radius,
                          particle_orient,
                          site,
                          rep_geom_params,
@@ -1331,7 +1331,7 @@ void model::DEMModel::createParticlesFromFile(
     // create transform
     auto p_transform = particle::ParticleTransform(
             site, util::Point(0., 0., 1.), particle_orient,
-            particle_radius / ref_p->getParticleRadius());
+            scale); //particle_radius / ref_p->getParticleRadius());
 
     if (p_transform.d_scale < 1.E-8) {
       std::cerr << "Error: check scale in transform. "
@@ -1374,7 +1374,7 @@ void model::DEMModel::createParticlesFromFile(
   }
 }
 
-void model::DEMModel::createGeometryAtSite(const double &particle_radius,
+double model::DEMModel::createGeometryAtSite(const double &particle_radius,
                                            const double &particle_orient,
                                            const util::Point &site,
                                            const std::vector<double> &rep_geom_params,
@@ -1383,6 +1383,9 @@ void model::DEMModel::createGeometryAtSite(const double &particle_radius,
   std::vector<double> params;
   for (auto x : rep_geom_params)
     params.push_back(x);
+
+  // scale to be used in transforming the geometry of representative particle to get this particle
+  double scale = 1.;
 
   if (util::methods::isTagInList(rep_geom_p->d_name,
                                  util::geometry::acceptable_geometries)) {
@@ -1401,6 +1404,8 @@ void model::DEMModel::createGeometryAtSite(const double &particle_radius,
       params[0] = particle_radius;
       for (int dof = 0; dof < 3; dof++)
         params[dof + 1] = site[dof];
+
+      scale = particle_radius / rep_geom_params[0];
     }
     else if (rep_geom_p->d_name == "drum2d") {
 
@@ -1410,10 +1415,12 @@ void model::DEMModel::createGeometryAtSite(const double &particle_radius,
       if (params.size() < num_params)
         params.resize(num_params);
 
-      params[0] = particle_radius; // biggger length along x-direction
+      params[0] = particle_radius; // bigger length along x-direction
       params[1] = particle_radius * rep_geom_params[1] / rep_geom_params[0]; // neck length along x-direction
       for (int dof = 0; dof < 3; dof++)
         params[dof + 2] = site[dof];
+
+      scale = params[0] / rep_geom_params[0];
     }
     else if (rep_geom_p->d_name == "rectangle") {
 
@@ -1427,6 +1434,8 @@ void model::DEMModel::createGeometryAtSite(const double &particle_radius,
       params[1] = particle_radius * rep_geom_params[1] / rep_geom_params[0]; // length along y-direction
       for (int dof = 0; dof < 3; dof++)
         params[dof + 2] = site[dof];
+
+      scale = params[0] / rep_geom_params[0];
     }
     else if (rep_geom_p->d_name == "cuboid") {
 
@@ -1435,11 +1444,24 @@ void model::DEMModel::createGeometryAtSite(const double &particle_radius,
       if (params.size() < 6)
         params.resize(6);
 
-      params[0] = particle_radius; // length is x-direction
-      params[1] = particle_radius * rep_geom_params[1] / rep_geom_params[0]; // length in y-direction
[…]
