### 2025-02-24 prashjha/PeriDEM#37 @nrichart
Hello @prashjha ,

I followed the instruction to run the examples in `PeriDEM/two_particles/twop_wall_concave_diff_material_diff_size/inp`
But there are no `input_8.yml`. I had to launch the python script that is mentioned to  modify the parameters.
You might need to add a step to run this script in the instructions or add some initial files.
Same thing with the compression example.
Once the files generates the examples run fine.

For the documentation the markdown version works fine, the links are pointing to valid points so for me it is fine. It is a good reading experience.

Thank you for all the modifications you have made.
