### 2025-02-24 #7525 @nrichart
My comments on the code where mostly addressed the remaining remarks I have a re a bit nite picking.

I just have a remark on the paper, there are no comparison to other work and how this work compares.
