+++ search/all_7.js
@@ -1,10 +1,0 @@
-var searchData=
-[
-  ['h_0',['h',['../namespaceparticle__wall.html#a928cfa4914ddbebc3278ad7f77e540d8',1,'particle_wall']]],
-  ['hatfunction_1',['hatFunction',['../namespaceutil.html#a51adb530fc272317c3a859af27787b1a',1,'util']]],
-  ['hatfunctionquick_2',['hatFunctionQuick',['../namespaceutil.html#a7ad8c95bec3f5405435dff75422e6993',1,'util']]],
-  ['help_3',['help',['../index.html#autotoc_md66',1,'Ask for help'],['../namespaceparticle__wall.html#a1da122babcef1b91f502d7b473b9daef',1,'particle_wall.help']]],
-  ['hexagon_4',['hexagon',['../classutil_1_1geometry_1_1Hexagon.html',1,'util::geometry::Hexagon'],['../classutil_1_1geometry_1_1Hexagon.html#a0b7179b6eac1d4aa4cc9aa10a9cdb5d3',1,'util::geometry::Hexagon::Hexagon()'],['../classutil_1_1geometry_1_1Hexagon.html#a87fcb9a61198f2194e1090757efa8f38',1,'util::geometry::Hexagon::Hexagon(double r, util::Point x=util::Point(0., 0., 0.), util::Point a=util::Point(1., 0., 0.), std::string description=&quot;&quot;)']]],
-  ['high_20fidelity_20modeling_20of_20granular_20media_20consisting_20of_20deformable_20complex_20shaped_20particles_5',['PeriDEM - High-fidelity modeling of granular media consisting of deformable complex-shaped particles',['../index.html',1,'']]],
-  ['horizon_6',['horizon',['../namespaceparticle__wall.html#aa08c5c90a5cffb178f26216ae2205144',1,'particle_wall']]]
-];
