+++ search/variables_1.js
@@ -1,5 +1,0 @@
-var searchData=
-[
-  ['beta_5fn_5feps_0',['beta_n_eps',['../namespaceparticle__wall.html#a6dbd3d0520a3383bb077521b34df86f8',1,'particle_wall']]],
-  ['beta_5fn_5ffactor_1',['beta_n_factor',['../namespaceparticle__wall.html#a21e03f959a50b7c6b99dfaafd0e9808e',1,'particle_wall']]]
-];

[…]
