+++ test/test_exec/meshpartitioning/testMeshPartitioning.cpp
@@ -1,0 +1,108 @@
+/*
+ * -------------------------------------------
+ * Copyright (c) 2021 - 2024 Prashant K. Jha
+ * -------------------------------------------
+ * PeriDEM https://github.com/prashjha/PeriDEM
+ *
+ * Distributed under the Boost Software License, Version 1.0. (See accompanying
+ * file LICENSE)
+ */
+
+#include "testMeshPartitioningLib.h"
+#include <PeriDEMConfig.h>
+#include <iostream>
+#include "util/io.h"                            // InputParser class
+#include "util/parallelUtil.h"                       // MPI-related functions
+#include <fmt/format.h>
+
+int main(int argc, char *argv[]) {
+
+  // init parallel
+  util::parallel::initMpi(argc, argv);
+  int mpiSize = util::parallel::mpiSize(), mpiRank = util::parallel::mpiRank();
+  util::io::print(fmt::format("Initialized MPI. MPI size = {}, MPI rank = {}\n", mpiSize, mpiRank));
+  util::io::print(util::parallel::getMpiStatus()->printStr());
+
+  util::io::InputParser input(argc, argv);
+
+  if (input.cmdOptionExists("-h") or !input.cmdOptionExists("-o")) {
+    // print help
+    std::cout << argv[0] << " (Version " << MAJOR_VERSION << "."
+              << MINOR_VERSION << "." << UPDATE_VERSION
+              << ") -o <0 to perform basic test, 1 on uniform mesh and 2 to test on user mesh> -n <grid-size> -p <number-partitions> -m <horizon-integer-factor> -f <mesh-filename>" << std::endl;
+    std::cout << "To perform basic test" << std::endl;
+    std::cout << argv[0] << " -o 0" << std::endl;
+    std::cout << "To test on uniform mesh" << std::endl;
+    std::cout << argv[0] << " -o 1 -p 4 -m 4 -n 10 " << std::endl;
+    std::cout << "To test on user-provided mesh (filename = filepath/meshfile.vtu)" << std::endl;
+    std::cout << argv[0] << " -o 2 -p 4 -m 4 -f filepath/meshfile.vtu" << std::endl;
+    exit(EXIT_FAILURE);
+  }
+
+  // read input
+  size_t testOption, nGrid(0), nPart, mHorizon;
+  std::string meshFilename("");
+
+  if (input.cmdOptionExists("-o")) testOption = size_t(std::stoi(input.getCmdOption("-o")));
+  else {
+    std::cerr << "Test requires following arguments -o <option integer> -n <grid-size> -p <number-partition> -m <horizon-integer-factor> -f <mesh-filename>\n";
+    exit(1);
+  }
+
+  if (testOption == 0) {
+    std::cout << "testMeshPartitioning: Simple test of metis graph partitioning\n\n";
+    test::testGraphPartitioningSimple();
+  } else if (testOption == 1 or testOption == 2) {
+
+    if (input.cmdOptionExists("-n"))
+      nGrid = size_t(std::stoi(input.getCmdOption("-n")));
+    else {
+      // set only if we are running test using in-built mesh
+      if (testOption == 1) {
+        nGrid = 50;
+        std::cout << "Running test with default grid size = "
+                  << nGrid << std::endl;
+      }
+    }
+
+    if (input.cmdOptionExists("-p"))
+      nPart = size_t(std::stoi(input.getCmdOption("-p")));
+    else {
+      nPart = 4;
+      std::cout << "Running test with default number of partitions = "
+                << nPart << std::endl;
+
+    }
+
+    if (input.cmdOptionExists("-m"))
+      mHorizon = size_t(std::stoi(input.getCmdOption("-m")));
+    else {
+      mHorizon = 4;
[…]
