### 2025-04-16 prashjha/PeriDEM#37 @prashjha
Hi @nrichart, Thank you for the comments. It is crazy busy for me, but I am sorry for not addressing it sooner. I provide the `run.sh` file in [examples/PeriDEM/two_particles/twop_wall_concave_diff_material_diff_size](https://github.com/prashjha/PeriDEM/tree/main/examples/PeriDEM/two_particles/twop_wall_concave_diff_material_diff_size), which first calls the Python script to generate an input file and then runs PeriDEM. A README.md file in the examples directory mentions a Python script. I think this should be enough information.

Thank you for your comments! I am closing the issue. But if you are not happy, just reopen it again and let me know!
