### 2025-04-22 #7525 @nrichart
@prashjha thank you for your work on the code and the documentation.
In my last message here I just also had a remark on the paper, I was wondering how your work compares to other works.
In particular in the review check list we have:
>  - **A statement of need**: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
>  - **State of the field**: Do the authors describe how this software compares to other commonly-used packages?

And for me your paper is missing this comparison to other work. When you say pure DEM struggles to represent granular media with large deformation, do you have any reference.
