### 2025-05-12 #7525 @matthewfeickert
@prashjha Can you address/respond to @nrichart's question in https://github.com/openjournals/joss-reviews/issues/7525#issuecomment-2820584112 (either here or in an GitHub Issue on https://github.com/prashjha/PeriDEM)?
