### 2025-05-22 1dced50
WIP: provide key methods to easily setup particle simulation in c++ file itself.
