+++ src/inp/bcDeck.h
@@ -1,0 +1,199 @@
+/*
+* -------------------------------------------
+ * Copyright (c) 2021 - 2024 Prashant K. Jha
+ * -------------------------------------------
+ * PeriDEM https://github.com/prashjha/PeriDEM
+ *
+ * Distributed under the Boost Software License, Version 1.0. (See accompanying
+ * file LICENSE)
+ */
+
+#ifndef INP_BCDECK_H
+#define INP_BCDECK_H
+
+#include <string>
+
+#include "bcBaseDeck.h"
+#include "util/io.h"
+
+namespace inp {
+  /**
+   * \ingroup Input
+   */
+  /**@{*/
+
+  /*! @brief Structure to read and store test-related input data */
+  struct BCDeck {
+
+    /*! @brief Force loading deck */
+    std::vector<inp::BCBaseDeck> d_forceDeck;
+
+    /*! @brief Displacement loading deck */
+    std::vector<inp::BCBaseDeck> d_dispDeck;
+
+    /*! @brief Initial condition deck */
+    std::vector<inp::BCBaseDeck> d_icDeck;
+
+    /*! @brief Gravity loading value */
+    bool d_gravityActive;
+
+    /*! @brief Gravity loading value */
+    util::Point d_gravity;
+
+
+    /*!
+     * @brief Constructor
+     */
+    BCDeck(const json &j = json({}))
+      : d_gravityActive(false) {
+      readFromJson(j);
+    };
+
+    /*!
+     * @brief Constructor
+     */
+    BCDeck(size_t nForceSets, size_t nDispSets = 0, size_t nICSets = 0,
+           bool gravityActive = false, util::Point gravity = util::Point())
+        : d_gravityActive(gravityActive), d_gravity(gravity) {
+
+        d_forceDeck.resize(nForceSets);
+        d_dispDeck.resize(nDispSets);
+        d_icDeck.resize(nICSets);
+    };
+
+    /*!
+     * @brief Returns example JSON object for ModelDeck configuration
+     * @return JSON object with example configuration
+     */
+    static json getExampleJson(size_t nForceSets = 0, size_t nDispSets = 0, size_t nICSets = 0,
+          bool gravityActive = false, util::Point gravity = util::Point()) {
+
+      if (nForceSets + nDispSets + nICSets == 0 and !gravityActive)
+        return json({});
+
+      auto j = json({});
+
+      if (gravityActive) j["Force_BC"] = {{"Gravity", gravity.toVec()}};
+
+      if (nForceSets > 0) {
+        j["Force_BC"] = {{"Sets", nForceSets}};
+        // add empty objects for each set
+        //         for (size_t i = 0; i < nForceSets; i++) j.at("Force_BC")["Set_" + std::to_string(i + 1)] = json({});
+      }
+
+      if (nDispSets > 0) {
+        j["Displacement_BC"] = {{"Sets", nDispSets}};
+        // add empty objects for each set
+        //        for (size_t i = 0; i < nDispSets; i++)
+        //          j.at("Displacement_BC") = json{{"Set_" + std::to_string(i + 1), ""}};
+      }
+
+      if (nICSets > 0) {
+        j["IC"] = {{"Sets", nICSets}};
+        // add empty objects for each set
+        //        for (size_t i = 0; i < nICSets; i++)
+        //          j.at("IC") = json{{"Set_" + std::to_string(i + 1), ""}};
+      }
+
+      return j;
+    }
+
+    /*!
+     * @brief Reads from json object
+     */
+    void readFromJson(const json &j) {
+
+      if (j.empty())
+        return;
+
+      // gravity
+      if (j.find("Force_BC") != j.end()) {
+        if (j.at("Force_BC").find("Gravity") != j.at("Force_BC").end()) {
+          d_gravityActive = true;
[…]
