+++ src/inp/modelDeck.h
@@ -1,0 +1,217 @@
+/*
+ * -------------------------------------------
+ * Copyright (c) 2021 - 2024 Prashant K. Jha
+ * -------------------------------------------
+ * PeriDEM https://github.com/prashjha/PeriDEM
+ *
+ * Distributed under the Boost Software License, Version 1.0. (See accompanying
+ * file LICENSE)
+ */
+
+#ifndef INP_MODELDECK_H
+#define INP_MODELDECK_H
+
+#include <string>
+#include "util/io.h"
+
+namespace inp {
+  /**
+   * \ingroup Input
+   */
+  /**@{*/
+
+  /*! @brief Structure to read and store model related input data */
+  struct ModelDeck {
+    /*!
+     * @brief Simulation type
+     *
+     * List of allowed values are:
+     * - \a explicit
+     * - \a implicit
+     */
+    std::string d_simType;
+
+    /*! @brief Flag indicating if this restart problem */
+    bool d_isRestartActive;
+
+    /*!
+     * @brief Tag for spatial discretization
+     *
+     * List of allowed values are:
+     * - \a **finite_difference**
+     * - \a **weak_finite_element**
+     * - \a **nodal_finite_element**
+     * - \a **truss_finite_element**
+     */
+    std::string d_spatialDiscretization;
+
+    /*!
+     * @brief Tag for time discretization
+     *
+     * List of allowed values are:
+     * - *empty string*
+     * - \a **central_difference**
+     * - \a **velocity_verlet**
+     */
+    std::string d_timeDiscretization;
+
+    /*! @brief Flag to indicate if we should populate element-node connectivity data in meshes */
+    bool d_populateElementNodeConnectivity;
+
+    /*! @brief Order of quadrature approximation for strain and stress computation (default is 1) */
+    size_t d_quadOrder;
+
+    /*! @brief Specify if this is single or multi particle simulation
+     * Expected value is either 'Single_Particle' or 'Multi_Particle'.
+     *
+     * This flag is used to populate input deck data. For the case when
+     * we consider single particle and its deformation, one do not have to specify data such as zones and contact.
+     * */
+    std::string d_particleSimType;
+
+    /*! @brief Dimension */
+    size_t d_dim;
+
+    /*! @brief Final simulation time */
+    double d_tFinal;
+
+    /*! @brief Size of time steps */
+    double d_dt;
+
+    /*! @brief Number of time steps */
+    size_t d_Nt;
+
+    /*! @brief Seed for random calculations (if any) */
+    int d_seed;
+
+    /*!
+     * @brief Constructor
+     */
+    ModelDeck(const json &j = json({}))
+        : d_dim(0), d_isRestartActive(false), d_populateElementNodeConnectivity(false),
+          d_tFinal(0.), d_dt(0.), d_Nt(0),
+          d_particleSimType(""), d_seed(0), d_quadOrder(1) {
+
+      readFromJson(j);
+    };
+
+    ModelDeck(size_t dim, double tFinal = 1.0, size_t Nt = 10,
+              std::string spatialDiscretization = "finite_difference",
+              std::string timeDiscretization = "central_difference",
+              bool populateElementNodeConnectivity = true,
+              size_t quadOrder = 2,
+              std::string particleSimType = "Multi_Particle",
+              int seed = 0)
+        : d_dim(dim), d_isRestartActive(false),
+          d_spatialDiscretization(spatialDiscretization),
+          d_timeDiscretization(timeDiscretization),
+          d_populateElementNodeConnectivity(populateElementNodeConnectivity),
+          d_tFinal(tFinal), d_dt(0.), d_Nt(Nt),
+          d_particleSimType(particleSimType), d_seed(seed), d_quadOrder(quadOrder) {
+
[…]
