+++ test/test_exec/particle/testParticle.cpp
@@ -1,0 +1,27 @@
+/*
+ * -------------------------------------------
+ * Copyright (c) 2021 - 2024 Prashant K. Jha
+ * -------------------------------------------
+ * PeriDEM https://github.com/prashjha/PeriDEM
+ *
+ * Distributed under the Boost Software License, Version 1.0. (See accompanying
+ * file LICENSE)
+ */
+
+#include "testParticleLib.h"
+#include "util/io.h"
+#include "util/parallelUtil.h"                       // MPI-related functions
+#include <fmt/format.h>
+
+int main() {
+
+  // init parallel
+  util::parallel::initMpi();
+  int mpiSize = util::parallel::mpiSize(), mpiRank = util::parallel::mpiRank();
+  util::io::print(fmt::format("Initialized MPI. MPI size = {}, MPI rank = {}\n", mpiSize, mpiRank));
+  util::io::print(util::parallel::getMpiStatus()->printStr());
+
+  test::testTransform();
+
+  return EXIT_SUCCESS;
+}

[…]
