### 2025-05-25 7c0021d
WIP: modular library. Progress: added mesh_gen folder to define Gmsh-based mesh generation for particles; better organization of libraries; test/test_data/peridem/twop_circ/main.cpp implements twop_circ test without needing any input file. Our main goal is to minimize the complex input file preparation. We want to define the problem in a main.cpp file without less efforts. Everytime, we want to create a new simulation, we can either define main.cpp or reuse json file.
