### 2025-06-03 68eb2cf
add python mesh functions using gmsh. Special attention is given to geometries for which we can generate symmetric mesh. This is crucial in certain applications to avoid artificial effects due non uniform mesh.
