+++ tools/python_utils/geom_util.py
@@ -1,0 +1,381 @@
+import os
+import sys
+import numpy as np
+import gmsh
+import math
+
+def rotate(p, theta, axis):
+  """
+  Returns rotation of vector about specified axis by specified angle
+
+  Parameters
+  ----------
+  p: list
+      Coordinates of vector
+  theta : float
+      Angle of rotation
+  axis : list
+      Axis of rotation
+
+  Returns
+  -------
+  list
+      Coordinates of rotated vector
+  """
+  p_np, axis_np = np.array(p), np.array(axis)
+  ct, st = np.cos(theta), np.sin(theta)
+  p_dot_n = np.dot(p_np,axis_np)
+  n_cross_p = np.cross(axis_np, p_np)
+
+  return (1. - ct) * p_dot_n * axis_np + ct * p_np + st * n_cross_p
+
+def get_ref_rect_points(center, Lx, Ly, add_center = False):
+  """
+  Returns size points on reference rectangle
+
+  Reference rectangle:
+  
+  +-------------------+
+  |                   |
+  |                   |
+  |                   |
+  +-------------------+
+
+  Parameters
+  ----------
+  center: list
+      Coordinates of center of rectangle
+  Lx : float
+      Length of rectangle in x-direction
+  Ly : float
+      Length of rectangle in y-direction
+  add_center : bool
+      True if we include the center to the returned list (first point in the returned list will be the center if the flag is true)
+
+  Returns
+  -------
+  list
+      Coordinates of points
+  """
+
+  p1 = [center[0] - Lx/2., center[1] - Ly/2., center[2]]
+  p2 = [center[0] + Lx/2., center[1] - Ly/2., center[2]]
+  p3 = [center[0] + Lx/2., center[1] + Ly/2., center[2]]
+  p4 = [center[0] - Lx/2., center[1] + Ly/2., center[2]]
+
+  return [p1, p2, p3, p4] if not add_center else [center, p1, p2, p3, p4]
+
+def get_ref_triangle_points(center, radius, add_center = False):
+  """
+  Returns size points on reference triangle
+
+  Reference triangle:
+              
+      + v2
+      | \
+      |      \
+      |            \  
+      |    o           + v1
+      |            /
+      |      /
+      | /
+      + v3
+  
+  Radius: o-v1 
+  Axis vector: o-v1
+  Rotation axis: [0, 0, 1]
+  
+  Parameters
+  ----------
+  center: list
+      Coordinates of center of triangle
+  radius : float
+      Radius of triangle
+  add_center : bool
+      True if we include the center to the returned list (first point in the returned list will be the center if the flag is true)
+
+  Returns
+  -------
+  list
+      Coordinates of points
+  """
+  axis = [1., 0., 0.]
+  rotate_axis = [0., 0., 1.]
+
+  points = []
+  if add_center:
+    points.append(center)
+  
+  for i in range(3):
+    xi = rotate(axis, i*2*np.pi/3., rotate_axis)
+    points.append([center[i] + radius * xi[i] for i in range(3)])
+
+  return points
+
+def get_ref_hex_points(center, radius, add_center = False):
+  """
+  Returns size points on reference hexagon
+
+  Reference hexagon:
+  
+                          v3           v2
+                           +           +
+  
+  
+                        +         o           +
+                       v4         x            v1
+  
+                           +           +
+                          v5           v6
+  
+  Radius: x-v1
+  Axis vector: x-v1
+  Rotation axis: [0, 0, 1]
+
+  Parameters
+  ----------
+  center: list
+      Coordinates of center of hexagon
+  radius : float
+      Radius of hexagon
+  add_center : bool
[…]
