+++ search/all_d.js
@@ -1,22 +1,0 @@
-var searchData=
-[
-  ['of_20contents_0',['Table of contents',['../index.html#autotoc_md47',1,'']]],
-  ['of_20granular_20media_20consisting_20of_20deformable_20complex_20shaped_20particles_1',['PeriDEM - High-fidelity modeling of granular media consisting of deformable complex-shaped particles',['../index.html',1,'']]],
-  ['of_20the_20code_2',['Compute times for various examples (From old version of the code!)',['../md__2home_2user_2project_2examples_2README.html#autotoc_md24',1,'']]],
-  ['old_20version_20of_20the_20code_3',['Compute times for various examples (From old version of the code!)',['../md__2home_2user_2project_2examples_2README.html#autotoc_md24',1,'']]],
-  ['on_20modifying_20input_20yaml_20file_4',['Important remark on modifying input.yaml file',['../md__2home_2user_2project_2examples_2README.html#autotoc_md22',1,'']]],
-  ['open_5',['open',['../classrw_1_1writer_1_1Writer.html#a99073ac82c4547629307fb269235d051',1,'rw::writer::Writer']]],
-  ['operator_28_29_6',['operator()',['../structutil_1_1Matrix3.html#a89cfa6b5806af9f106fb8e730b9ea380',1,'util::Matrix3::operator()()'],['../classutil_1_1DistributionSample.html#acaf45f46d05ea3ea27b62c0cf110b995',1,'util::DistributionSample::operator()()'],['../structutil_1_1SymMatrix3.html#a1a59be60694a1f09d5f499521c628472',1,'util::SymMatrix3::operator()(size_t i, size_t j) const'],['../structutil_1_1SymMatrix3.html#adc90321fb6490d53ee2e554a74865a2c',1,'util::SymMatrix3::operator()(size_t i, size_t j)'],['../structutil_1_1SymMatrix3.html#a770f6d01af7fb582ca11464cedbeebde',1,'util::SymMatrix3::operator()(size_t i) const'],['../structutil_1_1SymMatrix3.html#af753b40ad0442741692f00419b352cbd',1,'util::SymMatrix3::operator()(size_t i)'],['../structutil_1_1Matrix3.html#a615758cd442f2e4a9ca2e9c42d17af43',1,'util::Matrix3::operator()(size_t i, size_t j) const'],['../structutil_1_1Matrix3.html#a25dc047f28da7e8161608ad1e790e9fe',1,'util::Matrix3::operator()(size_t i, size_t j)'],['../structutil_1_1Matrix3.html#a4203eb8193d57736a8cbbf5af97824c4',1,'util::Matrix3::operator()(size_t i) const']]],
-  ['operator_2a_7',['operator*',['../structutil_1_1Point.html#ab33a890b402ee52391b8dba83231f8d7',1,'util::Point::operator*'],['../structutil_1_1Point.html#a24217bbdd160cd37b40451b39bb46620',1,'util::Point::operator*'],['../structutil_1_1Point.html#a7df34a1e89835339c5c0defc97a0b31a',1,'util::Point::operator*']]],
-  ['operator_2a_3d_8',['operator*=',['../structutil_1_1Point.html#a40a7aa8c025ee29760e55ec97410c594',1,'util::Point::operator*=(const double b)'],['../structutil_1_1Point.html#a2fa1f5c0b9f1b0944853619eddfa49ac',1,'util::Point::operator*=(const Point &amp;b)']]],
[…]
