+++ triElem_8h.js
@@ -1,4 +1,0 @@
-var triElem_8h =
-[
-    [ "fe::TriElem", "classfe_1_1TriElem.html", "classfe_1_1TriElem" ]
-];

[…]
