### 2025-06-07 bbc826b
add circle mesh generator to library. Note: included json conflicts with json shipped with vtk library. Easy fix is to have included json version match the json shipped with vtk in mac
