+++ src/util/json.h
@@ -1,0 +1,18 @@
+/*
+ * -------------------------------------------
+ * Copyright (c) 2021 - 2024 Prashant K. Jha
+ * -------------------------------------------
+ * PeriDEM https://github.com/prashjha/PeriDEM
+ *
+ * Distributed under the Boost Software License, Version 1.0. (See accompanying
+ * file LICENSE)
+ */
+
+#ifndef UTILS_UTIL_JSON_H
+#define UTILS_UTIL_JSON_H
+
+
+#include <nlohmann_json/json.hpp>
+using json = nlohmann::ordered_json;
+
+#endif // UTILS_UTIL_JSON_H
