### 2025-06-07 6d06b51
add ellipse geometry. It requires six parameters, rx, ry, x, y, z, theta
