+++ search/namespaces_2.js
@@ -1,6 +1,0 @@
-var searchData=
-[
-  ['discretize_5fcircle_0',['discretize_circle',['../namespacediscretize__circle.html',1,'']]],
-  ['discretize_5fcircle_5fdense_1',['discretize_circle_dense',['../namespacediscretize__circle__dense.html',1,'']]],
-  ['discretize_5fcircle_5fequidist_2',['discretize_circle_equidist',['../namespacediscretize__circle__equidist.html',1,'']]]
-];
