+++ navtreedata.js
@@ -1,0 +1,151 @@
+/*
+ @licstart  The following is the entire license notice for the JavaScript code in this file.
+
+ The MIT License (MIT)
+
+ Copyright (C) 1997-2020 by Dimitri van Heesch
+
+ Permission is hereby granted, free of charge, to any person obtaining a copy of this software
+ and associated documentation files (the "Software"), to deal in the Software without restriction,
+ including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ furnished to do so, subject to the following conditions:
+
+ The above copyright notice and this permission notice shall be included in all copies or
+ substantial portions of the Software.
+
+ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING
+ BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+ @licend  The above is the entire license notice for the JavaScript code in this file
+*/
+var NAVTREE =
+[
+  [ "PeriDEM", "index.html", [
+    [ "PeriDEM - High-fidelity modeling of granular media consisting of deformable complex-shaped particles", "index.html", "index" ],
+    [ "PeriDEM Examples", "md__2home_2user_2project_2examples_2PeriDEM_2README.html", [
+      [ "single_particle", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md1", [
+        [ "single_particle_circle", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md2", null ],
+        [ "single_particle_rectangle_inbuilt_mesh", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md3", null ]
+      ] ],
+      [ "two_particles", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md4", [
+        [ "twop_circ", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md5", null ],
+        [ "twop_concave_and_hex", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md6", null ],
+        [ "twop_wall_concave_diff_material_diff_size", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md7", null ]
+      ] ],
+      [ "compressive_test", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md8", [
+        [ "compression_small_set", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md9", null ],
+        [ "compression_large_set", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md10", null ]
+      ] ],
+      [ "attrition_tests", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md11", [
+        [ "attrition_mix_particles_small_set", "md__2home_2user_2project_2examples_2PeriDEM_2README.html#autotoc_md12", null ],
[…]
