### 2025-06-07 8f73cbd
add hanging node tests before writing mesh to file
