+++ tools/python_utils/gmsh_particles.py
@@ -8,6 +8,7 @@ from util import *
 from geom_util import *
 
 
+
 def circle_mesh_symmetric(xc = [0., 0., 0.], r = 1., h = 0.1, filename = 'mesh', vtk_out = False, symmetric_mesh = True):
   """
   Create a circular mesh, either symmetric or full.
@@ -77,6 +78,9 @@ def circle_mesh_symmetric(xc = [0., 0., 0.], r = 1., h = 0.1, filename = 'mesh',
 
     # generate mesh
     gmsh.model.mesh.generate(3)
+
+  # Check for hanging nodes before writing
+  check_hanging_nodes()
 
   # write
   gmsh.write(filename + '.msh')
@@ -157,6 +161,9 @@ def ellipse_mesh_symmetric(xc = [0., 0., 0.], rx = 1., ry = 0.5, h = 0.1, filena
 
     # generate mesh
     gmsh.model.mesh.generate(3)
+
+  # Check for hanging nodes before writing
+  check_hanging_nodes()
 
   # Write output files
   gmsh.write(filename + '.msh')
@@ -243,6 +250,9 @@ def sphere_mesh_symmetric(xc = [0., 0., 0.], r = 1., h = 0.1, filename = 'mesh',
     # Generate 3D mesh
     gmsh.model.mesh.generate(3)
   
+  # Check for hanging nodes before writing
+  check_hanging_nodes()
+
   # Write output files
   gmsh.write(filename + '.msh')
   if vtk_out:
@@ -328,6 +338,9 @@ def rectangle_mesh_symmetric(xc = [0., 0., 0.], Lx = 1., Ly = 1., h = 0.1, filen
     # generate mesh
     gmsh.model.mesh.generate(3)
 
+  # Check for hanging nodes before writing
+  check_hanging_nodes()
+
   # write
   gmsh.write(filename + '.msh')
   if vtk_out:
@@ -408,6 +421,9 @@ def hexagon_mesh_symmetric(xc = [0., 0., 0.], r = 1., h = 0.1, filename = 'mesh'
 
     # generate mesh
     gmsh.model.mesh.generate(3)
+
+  # Check for hanging nodes before writing
+  check_hanging_nodes()
 
   # Write output files
   gmsh.write(filename + '.msh')
@@ -497,6 +513,9 @@ def drum2d_mesh_symmetric(xc = [0., 0., 0.], r = 1., width = 1., h = 0.1, filena
     # generate mesh
     gmsh.model.mesh.generate(3)
 
+  # Check for hanging nodes before writing
+  check_hanging_nodes()
+
   # Write output files
   gmsh.write(filename + '.msh')
   if vtk_out:
@@ -565,6 +584,9 @@ def polygon_mesh_symmetric(points, theta, xc=[0., 0., 0.], h=0.1, filename='mesh
   
   # Translate to final position if needed
   gmsh_translate(xc)
+
+  # Check for hanging nodes before writing
+  check_hanging_nodes()
   
   # Write output files
   gmsh.write(filename + '.msh')
@@ -636,9 +658,17 @@ def cylindrical2d_wall_mesh(center=[0., 0., 0.], outer_radius=1.0, inner_radius=
   # Create surface with hole
   s1 = gmsh.model.geo.addPlaneSurface([outer_loop, inner_loop])
   
-  # Synchronize and generate mesh
+  # Synchronize geometry before adding physical groups
   gmsh.model.geo.synchronize()
+  
+  # Add physical group for the surface
+  gmsh.model.addPhysicalGroup(2, [s1], 1)
+  
+  # Generate mesh
[…]
