+++ examples/PeriDEM/compressive_test/compression_large_set/inp/problem_setup.py
@@ -1,0 +1,1635 @@
+import os
+import numpy as np
+# import csv
+import sys
+# import pyvista as pv
+
+def print_bool(arg, prefix = ""):
+
+  str = prefix
+  if arg == True:
+    str += "True\n"
+  else:
+    str += "False\n"
+  return str
+
+def print_dbl(arg, prefix = ""):
+
+  str = prefix + "%4.6e\n" % (arg)
+  return str
+
+def print_int(arg, prefix = ""):
+  str = prefix + "%d\n" % (arg)
+  return str
+
+def print_dbl_list(arg, prefix = ""):
+  str = prefix + "["
+  N = len(arg)
+  for i in range(N):
+    str += "%4.6e" % (arg[i])
+    if i < N - 1:
+      str += ", "
+    else:
+      str += "]\n"
+
+  return str
+
+def print_int_list(arg, prefix = ""):
+  str = prefix + "["
+  N = len(arg)
+  for i in range(N):
+    str += "%d" % (arg[i])
+    if i < N - 1:
+      str += ", "
+    else:
+      str += "]\n"
+
+  return str
+
+def serialize_matrix_list(p):
+  s = []
+  for q in p:
+    for w in q:
+      s.append(w)
+  
+  return s
+
+
+def write_point_geo(geof, p_id, x, h):
+    geof.write("Point(%d) = {%4.6e, %4.6e, %4.6e, %4.6e};\n" % (p_id, x[0], x[1], x[2], h))
+    return p_id + 1
+
+def write_line_geo(geof, l_id, p1_id, p2_id):
+    geof.write("Line(%d) = {%d, %d};\n" % (l_id, p1_id, p2_id))
+    return l_id + 1
+
+def write_cir_line_geo(geof, l_id, p1_id, p2_id, p3_id):
+    geof.write("Circle(%d) = {%d, %d, %d};\n" % (l_id, p1_id, p2_id, p3_id))
+    return l_id + 1
+
+def write_contact_zone_part(inpf, R_contact_factor, damping_active, friction_active, beta_n_eps, friction_coeff, Kn_factor, beta_n_factor, zone_string, Kn):
+  inpf.write("  Zone_%s:\n" % (zone_string))
+  inpf.write("    Contact_Radius_Factor: %4.6e\n" % (R_contact_factor))
+  
+  if damping_active == False:
+    inpf.write("    Damping_On: false\n")
+  if friction_active == False:
+    inpf.write("    Friction_On: false\n")
+
+  inpf.write("    Kn: %4.6e\n" % (Kn))
+  inpf.write("    Epsilon: %4.6e\n" % (beta_n_eps))
+  inpf.write("    Friction_Coeff: %4.6e\n" % (friction_coeff))
+  inpf.write("    Kn_Factor: %4.6e\n" % (Kn_factor))
+  inpf.write("    Beta_n_Factor: %4.6e\n" % (beta_n_factor))
+
+def write_material_zone_part(inpf, zone_string, horizon, rho, K, G, Gc):
+  inpf.write("  Zone_%s:\n" % (zone_string))
+  inpf.write("    Type: PDState\n")
+  inpf.write("    Horizon: %4.6e\n" % (horizon))
+  inpf.write("    Density: %4.6e\n" % (rho))
+  inpf.write("    Compute_From_Classical: true\n")
+  inpf.write("    K: %4.6e\n" % (K))
+  inpf.write("    G: %4.6e\n" % (G))
+  inpf.write("    Gc: %4.6e\n" % (Gc))
+  inpf.write("    Influence_Function:\n")
+  inpf.write("      Type: 1\n")
+
[…]
