### 2025-08-12 #7525 @nrichart
@prashjha thank you for your reply. I am not criticizing the content of your paper for say, more the form.

For me when I read in the Statement of need: `and its relation to other work?` and in the State of the field: `how this software compares to other commonly-used packages?`. I am hoping to see citation to other works either paper or codes even if they are not open-source. But I agree that this is my interpretation of these guidelines.
