### 2025-11-10 #7525 @matthewfeickert
@nrichart It looks like all of your issues opened on https://github.com/prashjha/PeriDEM are closed now. It seems that your review checklist https://github.com/openjournals/joss-reviews/issues/7525#issuecomment-2527390663 still has some areas unresolved though. The last discussion point between you and @prashjha that I see is on 2025-08-12 on https://github.com/openjournals/joss-reviews/issues/7525#issuecomment-3177822780 and https://github.com/openjournals/joss-reviews/issues/7525#issuecomment-3177894144 RE: references.

Can you please update us with the state of your review after this discussion?
