### 2025-11-10 prashjha/PeriDEM#45 @matthewfeickert
[JOSS Review]: Consider making build installable
With regards to https://github.com/openjournals/joss-reviews/issues/7525#issuecomment-3509070229, a suggested revision:

The current code seems to make strong assumptions about the user being on a Ubuntu release, which should not be required as system level packages can be usefully recommended but build requirements should be scoped to the individual software level, not their distributions for different flavors of Linux. This was brought up in PR https://github.com/prashjha/PeriDEM/pull/39 and addressed in https://github.com/prashjha/PeriDEM/blob/8f73cbd12fbf8d483bb46670661f9f2641d03c37/tools/README.md.

That is all helpful, but the following Pixi manifest (in the top level of the repository)

```toml
[workspace]
channels = ["conda-forge"]
name = "PeriDEM"
platforms = ["linux-64"]
version = "0.1.0"

# Currently can't install, only build?
# c.f. https://github.com/prashjha/PeriDEM/blob/8f73cbd12fbf8d483bb46670661f9f2641d03c37/README.md?plain=1#L264
[tasks.build]
cmd = """
rm -rf build && \
cmake   -DEnable_Documentation=OFF \
        -DEnable_Tests=ON \
        -DEnable_High_Load_Tests=OFF \
        -DDisable_Docker_MPI_Tests=ON \
        -DVTK_DIR="$CONDA_PREFIX" \
        -DMETIS_DIR="$CONDA_PREFIX" \
        -DCMAKE_BUILD_TYPE=Release \
        -DCMAKE_INSTALL_PREFIX="$CONDA_PREFIX" \
        -S . \
        -B build && \
cmake build -LH && \
cmake --build build --clean-first --parallel "$(nproc --ignore=2)" && \
ctest --verbose
"""

[dependencies]
cxx-compiler = ">=1.11.0,<2"
cmake = "<4"
make = ">=4.4.1,<5"
vtk = ">=9.5.2,<10"
yaml-cpp = ">=0.8.0,<0.9"
metis = ">=5.2.1,<6"
openmpi = ">=5.0.8,<6"
```

is sufficient (given the [current listed dependencies](https://github.com/prashjha/PeriDEM/blob/8f73cbd12fbf8d483bb46670661f9f2641d03c37/README.md?plain=1#L229)) to build the library and makes no Linux distribution assumptions beyond the computing platform

```console
$ pixi run build
...
[ 98%] Built target TestPeriDEM
[100%] Linking CXX executable bin/PeriDEM
[100%] Built target PeriDEM
```

though no tests are found(?)

```
...
[100%] Built target PeriDEM
UpdateCTestConfiguration  from :/tmp/PeriDEM/DartConfiguration.tcl
UpdateCTestConfiguration  from :/tmp/PeriDEM/DartConfiguration.tcl
Test project /tmp/PeriDEM
Constructing a list of tests
Updating test list for fixtures
Added 0 tests to meet fixture requirements
Checking test dependency graph...
Checking test dependency graph end
No tests were found!!!
```

so while I think your suggestions are good, I don't necessarily know if you want to tie yourself to particular Linux distributions or package managers

https://github.com/prashjha/PeriDEM/blob/8f73cbd12fbf8d483bb46670661f9f2641d03c37/tools/README.md?plain=1#L6

You could probably free yourself up to just report what you do and make it explicit what you expect users to install for themselves.

It **would be** a boon in general though if the code were installable to arbitrary user choice of directory tree, which is seems it currently is not

https://github.com/prashjha/PeriDEM/blob/8f73cbd12fbf8d483bb46670661f9f2641d03c37/README.md#L264-L267

That would be great to see.

Also, using

https://github.com/prashjha/PeriDEM/blob/8f73cbd12fbf8d483bb46670661f9f2641d03c37/CMakeLists.txt#L9

is problematic as it forces the _exact_ CMake policy regardless of the CMake version installed, and disallows using features from newer CMake releases if PeriDEM was going to be built as part of another project. CMake `<3.10` have already having support dropped in CMake `v4.3`, and so it is possible that in the very near future users would be required to downgrade their CMake installation to even be able to build `PeriDEM`.

As mentioned in the [`cmake_minimum_required` docs](https://cmake.org/cmake/help/latest/command/cmake_minimum_required.html), defining a `<policy_max>`

```diff
$ git diff -- CMakeLists.txt
diff --git a/CMakeLists.txt b/CMakeLists.txt
index 44031a98..53e00b8a 100755
--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -6,7 +6,7 @@
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
 # file LICENSE)

-cmake_minimum_required(VERSION 3.10)
+cmake_minimum_required(VERSION 3.10...4.3)

 #----------------------------------#
 #<<<<<<<<<< Project name >>>>>>>>>>#
```

 allows for this to be avoided and incrementally tested, and pick up warnings in advance (e.g. the ones that are present here).
