### 2025-11-10 #7525 @matthewfeickert
> Should I withdraw the article? I think the software can be improved and made user-friendly and easier to test.

@prashjha while you're free to do as you would like, I think that you've already done the work here to create software provides real use benefit, so I see no obvious reason why you should withdraw to create better code. If more user-friendly is within the scope of work that you could deliver for the review, there's no issue with that happening and being brought in (e.g. Issue https://github.com/prashjha/PeriDEM/issues/45). Though I think that if it would be a considerable amount of effort from you that if there are no usability issues open from reviewers that you should feel free to make those improvements as you see fit after publication when you have the time to do so.
