### 2025-11-13 #7525 @matthewfeickert
> I re checked the branch `v0.2.1` since it the version considered in this submission and there is no `install` which I did not notice the first time.

@nrichart as the reviews can result in changes to the codebase that would result in a new release just prior to publication we don't expect that authors will make multiple releases during the review process just for the review. It is a good idea to start the review at the Git tag of the submission, but for further review when there has been iteration by the author please use the `HEAD` of their default branch (or wherever the development for review revision is taking place) for the rest of the review.

> When I did the review I tested on a linux laptop and on a linux HPC cluster to not have the default path for the dependencies.

Great and thanks. It is excellent to test for portability. 👍
