### 2025-11-22 #7525 @JINGHU-UT
@matthewfeickert @prashjha
I think this software is a very significant contribution to the Discrete Element Method (DEM) community. Using peridynamics to handle breakage and deformation for complex particles is a smart way to bridge the gap where traditional DEM struggles.

I gave the build process a try on both Ubuntu and macOS following the tutorial, and it worked perfectly. However, I was unable to install it on Windows Subsystem for Linux (WSL), likely due to environment setup difficulties. Since WSL is a common workflow for Windows users doing preliminary development, I suggest either troubleshooting the WSL build process or noting this limitation in the documentation.

One minor suggestion: While the paper mentions how critical granular modeling is for sectors like mining, manufacturing, and geotechnical engineering, I think the Statement of Need and Background could be strengthened by adding a few more comparable papers from these applied fields. Connecting the tool to existing literature in these industries would better demonstrate its practical relevance. For instance, the crushing of sand particles is currently a hot topic in computational geotechnics, and citing recent work in that area would greatly highlight the timeliness of this software.
