### 2025-11-22 #7525 @matthewfeickert
Thanks @JINGHU-UT. Can you please open up Issues on https://github.com/prashjha/PeriDEM for your different comments and suggestions? When doing so, please mention `openjournals/joss-reviews#7525` so that a link is created to this Issue thread (and I can keep an eye on what is happening).
