### 2025-11-25 #7525 @matthewfeickert
👋 @prashjha can you give us an update on your time availability over the next month to follow up on Issues https://github.com/prashjha/PeriDEM/issues with the reviewers?
