### 2025-11-26 d869655
JOSS review: update summary and statement of need sections based on the reviewer comments. Also, shorten the remaining sections keeping in mind JOSS type articles
