+++ docs/joss/paper.bib
@@ -1,3 +1,90 @@
+@article{Jha2021, doi = {10.21105/joss.03020}, url = {https://doi.org/10.21105/joss.03020}, year = {2021}, publisher = {The Open Journal}, volume = {6}, number = {65}, pages = {3020}, author = {Jha, Prashant K. and Diehl, Patrick}, title = {NLMech: Implementation of finite difference/meshfree discretization of nonlocal fracture models}, journal = {Journal of Open Source Software} }
+
+@article{dosta2024comparing,
+  title={Comparing open-source DEM frameworks for simulations of common bulk processes},
+  author={Dosta, Maksym and Andr{\'e}, Damien and Angelidakis, Vasileios and Caulk, Robert A and Celigueta, Miguel Angel and Chareyre, Bruno and Dietiker, J-F and Girardot, J and Govender, Nicolin and Hubert, C and others},
+  journal={Computer Physics Communications},
+  volume={296},
+  pages={109066},
+  year={2024},
+  publisher={Elsevier}
+}
+
+@article{zhang_2024_deme,
+title = {Chrono {DEM-Engine}: A Discrete Element Method dual-{GPU} simulator with customizable contact forces and element shape},
+journal = {Computer Physics Communications},
+volume = {300},
+pages = {109196},
+year = {2024},
+issn = {0010-4655},
+doi = {https://doi.org/10.1016/j.cpc.2024.109196},
+author = {Ruochun Zhang and Bonaventura Tagliafierro and Colin {Vanden Heuvel} and Shlok Sabarwal and Luning Bakke and Yulong Yue and Xin Wei and Radu Serban and Dan Negruţ},
+keywords = {Discrete Element Method, GPU computing, Physics-based simulation, Scientific package, BSD3 open-source},
+}
+
+@article{chen2023mechanical,
+  title={Mechanical behavior and particle crushing of irregular granular material under high pressure using discrete element method},
+  author={Chen, Qinglin and Li, Zugui and Dai, Zeyu and Wang, Xiaojun and Zhang, Chao and Zhang, Xuepeng},
+  journal={Scientific Reports},
+  volume={13},
+  number={1},
+  pages={7843},
+  year={2023},
+  publisher={Nature Publishing Group UK London}
+}
+
+@article{harmon2021modeling,
+  title={Modeling connected granular media: Particle bonding within the level set discrete element method},
+  author={Harmon, John M and Karapiperis, Konstantinos and Li, Liuchi and Moreland, Scott and Andrade, Jos{\'e} E},
+  journal={Computer Methods in Applied Mechanics and Engineering},
+  volume={373},
+  pages={113486},
+  year={2021},
+  publisher={Elsevier}
+}
+
+@article{littlewood2024peridigm,
+  title={The peridigm meshfree peridynamics code},
+  author={Littlewood, David J and Parks, Michael L and Foster, John T and Mitchell, John A and Diehl, Patrick},
+  journal={Journal of Peridynamics and Nonlocal Modeling},
+  volume={6},
+  number={1},
+  pages={118--148},
+  year={2024},
+  publisher={Springer}
+}
+
+@article{govender2016blaze,
+  title={Blaze-DEMGPU: Modular high performance DEM framework for the GPU architecture},
+  author={Govender, Nicolin and Wilke, Daniel N and Kok, Schalk},
+  journal={SoftwareX},
+  volume={5},
+  pages={62--66},
+  year={2016},
+  publisher={Elsevier}
+}
+
+@article{yade2021,
+  title        = {Yade Documentation (3rd edition)},
[…]
