+++ docs/joss/paper.md
@@ -5,7 +5,6 @@ tags:
   - Peridynamics
   - Discrete element method
   - Fracture
-  - Computational mechanics
   - Particle breakage
   - Particle interlocking
 authors:
@@ -17,18 +16,24 @@ affiliations:
    index: 1
 header-includes:
    - \usepackage{amsfonts,amssymb,amsmath}
-date: 17 October 2024
+date: 25 November 2025
 bibliography: paper.bib
 
 ---
 
 # Summary
 
-Granular materials are crucial in various sectors, including geotechnical, manufacturing, and mining. Predictive modeling of these materials under extensive loading becomes challenging due to the deformation and breakage of particles and the complex contact mechanisms between complex-shaped particles undergoing considerable deformation. Focusing on the scenarios when particle deformation and breakage are crucial, the PeriDEM model introduced in [@jha2021peridynamics] is implemented in the PeriDEM library. The underlying idea is that individual particles are modeled as deformable solids using peridynamics theory, and the contact between two deforming particles is applied locally at the contact region, allowing the modeling of complex-shaped particles. Integrating peridynamics within the discrete element method (DEM) provides a flexible, hybrid framework that handles the contact mechanics at the particle boundary while accounting for the internal material response, including deformation and fracture. This opens up new avenues for exploring the interactions in granular systems, including developing constitutive laws for phenomenological continuum models, understanding effective behavior when subjected to extensive loading, and the impact of particle shape on particle dynamics. 
+Accurate simulation of granular materials under extreme mechanical conditions-such as crushing, fracture, and large deformation—remains a significant challenge in geotechnical, manufacturing, and mining applications. Classical discrete element method (DEM) models typically treat particles as rigid or nearly rigid bodies, limiting their ability to capture internal deformation and fracture. The PeriDEM library, first introduced in [@jha2021peridynamics], addresses this limitation by modeling particles as deformable solids using peridynamics, a nonlocal continuum theory that naturally accommodates fracture and large deformation. Inter-particle contact is handled using DEM-inspired local laws, enabling realistic interaction between complex-shaped particles.
+
[…]
