### 2025-11-26 957e445
JOSS review: fix reference call
