### 2025-11-26 ca615fb
JOSS review: replace theory by referring to specific sections of JMPS article
