### 2025-11-26 106b007
JOSS review: refine text
