+++ docs/joss/paper.md
@@ -1,21 +1,21 @@
 ---
 title: 'PeriDEM -- High-fidelity modeling of granular media consisting of deformable complex-shaped particles'
 tags:
-  - Granular media
-  - Peridynamics
-  - Discrete element method
-  - Fracture
-  - Particle breakage
-  - Particle interlocking
+ - Granular media
+ - Peridynamics
+ - Discrete element method
+ - Fracture
+ - Particle breakage
+ - Particle interlocking
 authors:
-  - name: Prashant Kumar Jha
+ - name: Prashant Kumar Jha
     orcid: 0000-0003-2158-364X
     affiliation: 1
 affiliations:
  - name: Department of Mechanical Engineering, South Dakota School of Mines and Technology, Rapid City, SD 57701, USA
    index: 1
 header-includes:
-   - \usepackage{amsfonts,amssymb,amsmath}
+ - \usepackage{amsfonts,amssymb,amsmath}
 date: 25 November 2025
 bibliography: paper.bib
 
@@ -23,17 +23,17 @@ bibliography: paper.bib
 
 # Summary
 
-Accurate simulation of granular materials under extreme mechanical conditions-such as crushing, fracture, and large deformation—remains a significant challenge in geotechnical, manufacturing, and mining applications. Classical discrete element method (DEM) models typically treat particles as rigid or nearly rigid bodies, limiting their ability to capture internal deformation and fracture. The PeriDEM library, first introduced in [@jha2021peridynamics], addresses this limitation by modeling particles as deformable solids using peridynamics, a nonlocal continuum theory that naturally accommodates fracture and large deformation. Inter-particle contact is handled using DEM-inspired local laws, enabling realistic interaction between complex-shaped particles.
+Accurate simulation of granular materials under extreme mechanical conditions, such as crushing, fracture, and large deformation, remains a significant challenge in geotechnical, manufacturing, and mining applications. Classical discrete element method (DEM) models typically treat particles as rigid or nearly rigid bodies, limiting their ability to capture internal deformation and fracture. The PeriDEM library, first introduced in [@jha2021peridynamics], addresses this limitation by modeling particles as deformable solids using peridynamics, a nonlocal continuum theory that naturally accommodates fracture and significant deformation. Inter-particle contact is handled using DEM-inspired local laws, enabling realistic interaction between complex-shaped particles.
 
 Implemented in C++, PeriDEM is designed for extensibility and ease of deployment. It relies on a minimal set of external libraries, supports multithreaded execution, and includes demonstration examples involving compaction, fracture, and rotational dynamics. The framework facilitates granular-scale simulations, supports the development of constitutive models, and serves as a foundation for multi-fidelity coupling in real-world applications.
 
 # Statement of Need
 
-Granular materials play a central role in many engineered systems, but modeling their behavior under high loading, deformation, and fragmentation remains an open problem. Popular open-source DEM codes such as YADE [@yade2021], BlazeDEM [@govender2016blaze], Chrono DEM-Engine [@zhang_2024_deme], and LAMMPS [@THOMPSON2022108171] are widely used but typically treat particles as rigid, limiting their accuracy in scenarios involving internal deformation and breakage. A recent review by Dosta et al. [@dosta2024comparing] compares these libraries across a range of bulk processes. Meanwhile, peridynamics-based codes like Peridigm [@littlewood2024peridigm] and NLMech [@Jha2021] offer detailed fracture modeling but do not capture realistic particle contact mechanics or bulk granular dynamics.
[…]
