### 2025-11-26 7c564e8
JOSS review: add acknowledgement
