### 2025-11-26 48a7518
JOSS review: add acknowledgement
