### 2025-11-26 ba3cc77
JOSS review: add acknowledgement
