### 2025-11-26 a406c24
JOSS review: final refinement
