### 2025-11-26 e6d34c0
JOSS review: final refinement
