### 2025-11-26 499efd7
JOSS review: final refinement
