### 2025-11-26 a517a15
JOSS review: final refinement
