+++ docs/joss/paper.bib
@@ -1,3 +1,13 @@
+@article{jha2025nodal,
+  title={Nodal finite element approximation of peridynamics},
+  author={Jha, Prashant K and Diehl, Patrick and Lipton, Robert},
+  journal={Computer Methods in Applied Mechanics and Engineering},
+  volume={434},
+  pages={117519},
+  year={2025},
+  publisher={Elsevier}
+}
+
 @article{Jha2021, doi = {10.21105/joss.03020}, url = {https://doi.org/10.21105/joss.03020}, year = {2021}, publisher = {The Open Journal}, volume = {6}, number = {65}, pages = {3020}, author = {Jha, Prashant K. and Diehl, Patrick}, title = {NLMech: Implementation of finite difference/meshfree discretization of nonlocal fracture models}, journal = {Journal of Open Source Software} }
 
 @article{dosta2024comparing,
