### 2025-11-26 1a082d0
JOSS review: final refinement
