### 2025-11-26 b7a679a
JOSS review: final refinement
