+++ docs/joss/paper.bib
@@ -108,6 +108,7 @@ keywords = {Molecular dynamics, Materials modeling, Parallel algorithms, LAMMPS}
 }
 
 @article{jha2021peridynamics,
+  author = {Jha, Prashant K and Desai, Prathamesh S and Bhattacharya, Debdeep and Lipton, Robert},
   title = {Peridynamics-based discrete element method (PeriDEM) model of granular systems involving breakage of arbitrarily shaped particles},
   journal = {Journal of the Mechanics and Physics of Solids},
   volume = {151},
@@ -116,7 +117,6 @@ keywords = {Molecular dynamics, Materials modeling, Parallel algorithms, LAMMPS}
   issn = {0022-5096},
   doi = {10.1016/j.jmps.2021.104376},
   url = {https://doi.org/10.1016/j.jmps.2021.104376},
-  author = {Jha, Prashant K. and Desai, Prathamesh S. and Bhattacharya, Debdeep and Lipton, Robert},
   keywords = {Peridynamics, Discrete element method, Particle attrition, Particle interlocking, Particle breakage, Granular media, Fracture}
 }
 
