### 2025-11-26 6ebf6be
JOSS review: final refinement
