### 2025-11-26 74f760d
JOSS review: final refinement
