+++ docs/joss/paper.bib
@@ -107,8 +107,8 @@ keywords = {Molecular dynamics, Materials modeling, Parallel algorithms, LAMMPS}
   url = {https://doi.org/10.1680/geot.1979.29.1.47}
 }
 
-@article{Jha2021JMPS,
-  author = {Jha, Prashant K and Desai, Prathamesh S and Bhattacharya, Debdeep and Lipton, Robert},
+@article{jha2021peridynamics,
+  author = {Jha, Prashant K. and Desai, Prathamesh S. and Bhattacharya, Debdeep and Lipton, Robert},
   title = {Peridynamics-based discrete element method (PeriDEM) model of granular systems involving breakage of arbitrarily shaped particles},
   journal = {Journal of the Mechanics and Physics of Solids},
   volume = {151},
