### 2025-11-26 e054c54
JOSS review: final refinement
