+++ docs/joss/paper.bib
@@ -1,6 +1,6 @@
 @article{jha2025nodal,
   title={Nodal finite element approximation of peridynamics},
-  author={Jha, Prashant K and Diehl, Patrick and Lipton, Robert},
+  author={Jha, Prashant K. and Diehl, Patrick and Lipton, Robert},
   journal={Computer Methods in Applied Mechanics and Engineering},
   volume={434},
   pages={117519},
