### 2025-11-26 d5891f9
JOSS review: final refinement
