### 2025-11-26 61ec5d6
JOSS review: final refinement
