### 2025-11-26 536ae10
JOSS review: final refinement
