+++ circle__wall__input__using__symmetric__mesh_8py.js
@@ -1,6 +1,0 @@
-var circle__wall__input__using__symmetric__mesh_8py =
-[
-    [ "create_input_file", "circle__wall__input__using__symmetric__mesh_8py.html#af6fc8d30f4a74bb437c3b7fbd099f15e", null ],
-    [ "inp_dir", "circle__wall__input__using__symmetric__mesh_8py.html#a407e5a890ee0bf8a2f830178940b059f", null ],
-    [ "pp_tag", "circle__wall__input__using__symmetric__mesh_8py.html#ad13951e5b89c7c50a3f60960ac66af81", null ]
-];
