+++ search/variables_12.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['update_5fversion_0',['UPDATE_VERSION',['../PeriDEMConfig_8h.html#a48876dfce083f8387a5244b429c2a55e',1,'PeriDEMConfig.h']]]
-];
