+++ search/classes_f.js
@@ -1,7 +1,0 @@
-var searchData=
-[
-  ['rectangle_0',['Rectangle',['../classutil_1_1geometry_1_1Rectangle.html',1,'util::geometry']]],
-  ['refparticle_1',['RefParticle',['../classparticle_1_1RefParticle.html',1,'particle']]],
-  ['restartdeck_2',['RestartDeck',['../structinp_1_1RestartDeck.html',1,'inp']]],
-  ['rnpmaterial_3',['RnpMaterial',['../classmaterial_1_1RnpMaterial.html',1,'material']]]
-];

[…]
