### 2025-11-26 f4f43c5
automated deployment to GitHub Pages: 8cd382b7feaa7e725dfdc51c2a882ab9b340999e
