+++ src/PeriDEMConfig.h
@@ -1,0 +1,20 @@
+/*
+ * -------------------------------------------
+ * Copyright (c) 2021 - 2024 Prashant K. Jha
+ * -------------------------------------------
+ * PeriDEM https://github.com/prashjha/PeriDEM
+ *
+ * Distributed under the Boost Software License, Version 1.0. (See accompanying
+ * file LICENSE)
+ */
+
+#ifndef CONFIG_H
+#define CONFIG_H
+
+#include<cstddef>
+
+size_t const MAJOR_VERSION = 0;
+size_t const MINOR_VERSION = 2;
+size_t const UPDATE_VERSION = 0;
+
+#endif
