+++ src/model/dem/demModel.h
@@ -1,0 +1,242 @@
+/*
+ * -------------------------------------------
+ * Copyright (c) 2021 - 2024 Prashant K. Jha
+ * -------------------------------------------
+ * PeriDEM https://github.com/prashjha/PeriDEM
+ *
+ * Distributed under the Boost Software License, Version 1.0. (See accompanying
+ * file LICENSE)
+ */
+
+#ifndef MODEL_FASTDEMMODEL_H
+#define MODEL_FASTDEMMODEL_H
+
+#include "../modelData.h"
+
+namespace model {
+
+/**
+ * \defgroup Explicit Explicit
+ */
+/**@{*/
+
+/*! @brief A class for *discrete element particle simulation* with
+ * *peridynamic model*
+ *
+ * We consider *central difference* for time integration.
+ *
+ * This class acts as a holder of lower rank classes, such as set of
+ * particles, neighborlist, etc., and uses the methods and data of the
+ * lower rank classes to perform calculation.
+ */
+class DEMModel : public ModelData {
+
+public:
+  /*!
+   * @brief Constructor
+   * @param deck The input deck
+   */
+  explicit DEMModel(inp::Input *deck, std::string modelName = "DEMModel");
+
+  /*! @brief Name of the model for logging purposes (useful if other classes are built on top of this class) */
+  std::string d_name;
+
+  /*! @brief Prints message if any of these two conditions are true
+   * 1. if check_condition == true and dbg_lvl > priority
+   * OR
+   * 2. dbg_lvl > override_priority
+   *
+   * @param oss Stream
+   * @param priority Priority of message
+   * @param check_condition Specify whether condition for logging/printing message is to be checked
+   * @param override_priority Specify new priority
+   * @param screen_out Specify whether to print the message also to screen (using std::cout)
+   */
+  void log(std::ostringstream &oss, int priority = 0, bool check_condition = true, int override_priority = -1, bool screen_out = false);
+
+  /*! @brief Prints message if any of these two conditions are true
+   * 1. if check_condition == true and dbg_lvl > priority
+   * OR
+   * 2. dbg_lvl > override_priority
+   *
+   * @param str Message string
+   * @param priority Priority of message
+   * @param check_condition Specify whether condition for logging/printing message is to be checked
+   * @param override_priority Specify new priority
+   * @param screen_out Specify whether to print the message also to screen (using std::cout)
+   */
+  void log(const std::string &str, int priority = 0, bool check_condition = true, int override_priority = -1, bool screen_out = false);
+
+  /*!
+   * @brief Main driver to simulate
+   * @param deck Input deck
+   */
+  virtual void run(inp::Input *deck);
+
+  /**
+   * @name Methods to initialize and run model
+   */
+  /**@{*/
+
+  /*!
+   * @brief Restarts the simulation from previous state
+   * @param deck Input deck
+   */
+  virtual void restart(inp::Input *deck);
+
+  /*! @brief Initialize remaining data members */
+  virtual void init();
+
+  /*! @brief Closure operations */
+  virtual void close();
+
+  /** @}*/
+
+  /**
+   * @name Methods to implement explicit time integration
+   */
+  /**@{*/
+
+  /*! @brief Perform time integration */
+  virtual void integrate();
+
+  /*!
+   * @brief Performs one time step
+   *
+   * Depending on the time-stepping scheme specified in the input file, this
+   * will either call integrateCD or integrateVerlet.
+   */
+  virtual void integrateStep();
+
+  /*!
+   * @brief Perform time integration using central-difference scheme
+   *
+   * Central difference scheme
[…]
