### 2025-11-26 5d3d2f9
(auto) Paper PDF Draft
