+++ hierarchy.js
@@ -1,94 +1,0 @@
-var hierarchy =
-[
-    [ "fe::BaseElem", "classfe_1_1BaseElem.html", [
-      [ "fe::LineElem", "classfe_1_1LineElem.html", null ],
-      [ "fe::QuadElem", "classfe_1_1QuadElem.html", null ],
-      [ "fe::TetElem", "classfe_1_1TetElem.html", null ],
-      [ "fe::TriElem", "classfe_1_1TriElem.html", null ]
-    ] ],
-    [ "material::BaseInfluenceFn", "classmaterial_1_1BaseInfluenceFn.html", [
-      [ "material::ConstInfluenceFn", "classmaterial_1_1ConstInfluenceFn.html", null ],
-      [ "material::GaussianInfluenceFn", "classmaterial_1_1GaussianInfluenceFn.html", null ],
-      [ "material::LinearInfluenceFn", "classmaterial_1_1LinearInfluenceFn.html", null ]
-    ] ],
-    [ "nsearch::BaseNSearch", "classnsearch_1_1BaseNSearch.html", [
-      [ "nsearch::NFlannSearchKd< dim >", "classnsearch_1_1NFlannSearchKd.html", null ]
-    ] ],
-    [ "particle::BaseParticle", "classparticle_1_1BaseParticle.html", null ],
-    [ "util::geometry::BoxPartition", "structutil_1_1geometry_1_1BoxPartition.html", null ],
-    [ "inp::ContactDeck", "structinp_1_1ContactDeck.html", null ],
-    [ "inp::ContactPairDeck", "structinp_1_1ContactPairDeck.html", null ],
-    [ "util::DistributionSample< T >", "classutil_1_1DistributionSample.html", null ],
-    [ "geometry::Fracture", "classgeometry_1_1Fracture.html", null ],
-    [ "util::geometry::GeomData", "structutil_1_1geometry_1_1GeomData.html", null ],
-    [ "util::geometry::GeomObject", "classutil_1_1geometry_1_1GeomObject.html", [
-      [ "util::geometry::AnnulusGeomObject", "classutil_1_1geometry_1_1AnnulusGeomObject.html", null ],
-      [ "util::geometry::Circle", "classutil_1_1geometry_1_1Circle.html", null ],
-      [ "util::geometry::ComplexGeomObject", "classutil_1_1geometry_1_1ComplexGeomObject.html", null ],
-      [ "util::geometry::Cube", "classutil_1_1geometry_1_1Cube.html", null ],
-      [ "util::geometry::Cuboid", "classutil_1_1geometry_1_1Cuboid.html", null ],
-      [ "util::geometry::Cylinder", "classutil_1_1geometry_1_1Cylinder.html", null ],
-      [ "util::geometry::Drum2D", "classutil_1_1geometry_1_1Drum2D.html", null ],
-      [ "util::geometry::Ellipse", "classutil_1_1geometry_1_1Ellipse.html", null ],
-      [ "util::geometry::Hexagon", "classutil_1_1geometry_1_1Hexagon.html", null ],
-      [ "util::geometry::Line", "classutil_1_1geometry_1_1Line.html", null ],
-      [ "util::geometry::NullGeomObject", "classutil_1_1geometry_1_1NullGeomObject.html", null ],
-      [ "util::geometry::Rectangle", "classutil_1_1geometry_1_1Rectangle.html", null ],
-      [ "util::geometry::Sphere", "classutil_1_1geometry_1_1Sphere.html", null ],
-      [ "util::geometry::Square", "classutil_1_1geometry_1_1Square.html", null ],
[…]
