+++ geom__util_8py.js
@@ -1,15 +1,0 @@
-var geom__util_8py =
-[
-    [ "check_hanging_nodes", "geom__util_8py.html#a13f030c6d181d0210c230fc01e33331c", null ],
-    [ "does_intersect", "geom__util_8py.html#aecc5408bdc61be1b564793f404660dea", null ],
-    [ "does_intersect_rect", "geom__util_8py.html#a17e97edccd39949ba20d501c5bc1c417", null ],
-    [ "get_gmsh_entities", "geom__util_8py.html#a4cff2d2279e07cecdbff9c469cadc986", null ],
-    [ "get_ref_drum_points", "geom__util_8py.html#a110d13875687bf3abefe2ffe9564db4c", null ],
-    [ "get_ref_hex_points", "geom__util_8py.html#ac89d32108c84f9108df8ddbb2b5ea77f", null ],
-    [ "get_ref_rect_points", "geom__util_8py.html#a71c0c6ad61c7bf3c47c0133794489ad8", null ],
-    [ "get_ref_triangle_points", "geom__util_8py.html#a6212d7d95d932b464e795348ccaa1fb1", null ],
-    [ "gmsh_transform", "geom__util_8py.html#a4da92e283f0d3adaeec14f498c0f2c10", null ],
-    [ "gmsh_transform_general", "geom__util_8py.html#a6e4b10f501c8203a352d2390f68b6561", null ],
-    [ "gmsh_translate", "geom__util_8py.html#a0402fa3ab58c62403ff40506f3453665", null ],
-    [ "rotate", "geom__util_8py.html#acd65ced5bcfff51014a4b8dc6e6a935e", null ]
-];
