+++ dir_0468377997edb5eba1ba270cc0d457d4.js
@@ -1,6 +1,0 @@
-var dir_0468377997edb5eba1ba270cc0d457d4 =
-[
-    [ "compression_large_set", "dir_2c1263c588fa837354af022f3f604589.html", "dir_2c1263c588fa837354af022f3f604589" ],
-    [ "compression_small_set", "dir_8863324614beb3c53b5de75eb445cea7.html", "dir_8863324614beb3c53b5de75eb445cea7" ],
-    [ "n500_circ_hex", "dir_1bdae2d96021d5e34a24d9f135945246.html", "dir_1bdae2d96021d5e34a24d9f135945246" ]
-];
