### 2025-11-29 bab82db
JOSS review: pixi.toml file to build library following matthewfeickert comment #45. Cleaned README.md in tools and mentions that the scripts are outdated. Updated README.md to refer to pixi file
