+++ README.md
@@ -25,6 +25,7 @@
   - [Developers](#Developers)
 
 ## Introduction
+
 Implementation of the high-fidelity model of granular media that combines the advantages 
 of peridynamics and discrete element method (DEM). 
 The model has the following advantages over existing mechanical models for granular media:
@@ -48,14 +49,17 @@ We have created channels on various platforms:
   * Email us if interested in joining the workspace.
 
 ## Documentation
+
 [Doxygen generated documentation](https://prashjha.github.io/PeriDEM/) details functions and objects in the library. 
 
 ## Tutorial
+
 We explain the setting-up of simulations in further details in [tutorial](./tutorial/README.md). 
 We consider `two-particle` test setup with non-circular particles and `compressive-test` to 
 discuss the various aspects of simulations.
 
 ## Examples
+
 We next highlight some key examples. Further details are available in [examples/README.md](./examples/README.md). 
 
 ### Two-particle tests
@@ -75,6 +79,7 @@ We next highlight some key examples. Further details are available in [examples/
 | Concave particles |
 
 ### Compressive tests
+
 Setup for this test consists of 502 circular and hexagonal-shaped particles of varying 
 radius and orientation inside a rectangle container. The container's top wall is moving 
 downward at a prescribed speed, resulting in the compression of the particle system. 
@@ -91,6 +96,7 @@ yielding of the system. For more details, we refer to
 
 
 ### Attrition tests
+
 We consider mix of different particles in a rotating container. Particles considered include circular, triangular, hexagonal, and drum shaped. Particles come in large and small shapes (their sizes are purturbed randomly). In order to to introduce diversity of material properties, we considered large particles to be tougher compared to the smaller ones. Setup files are in [examples/PeriDEM/attrition_tests](./examples/PeriDEM/attrition_tests)
 
 |                                         <img src="./assets/attrition_test_sim1.gif" width="250">                                         |                                                                              <img src="./assets/attrition_test_sim2.gif" width="250">                                                                               | 
@@ -102,14 +108,17 @@ Complex container geometries can be considered as well. For example, the image b
 <img src="./examples/PeriDEM/attrition_tests/sim4_multi_particle_circ_tri_drum_hex_with_rotating_rectangle_container_with_protrusion_and_opening_within_bigger_rectangle_container/init_view.png" width="600">
 
 ### Single particle deformation
+
 We can use `PeriDEM` executable or `Peridynamics` executable in `apps` directory to simulate the deformation of single particle/structure using peridynamics. See [examples/README.md](./examples/README.md) and [examples/Peridynamics](./examples/Peridynamics) folder. 
 
 ## Brief implementation details
+
 The main implementation of the model is carried out in the model directory [dem](./src/model/dem). 
 The model is implemented in class [DEMModel](./src/model/dem/demModel.cpp). 
 Function `DEMModel::run()` performs the simulation. We next look at some key methods in `DEMModel` in more details:
 
 ### DEMModel::run()
+
 This function does three tasks:
 ```cpp
 void model::DEMModel::run(inp::Input *deck) {
@@ -129,6 +138,7 @@ In `DEMModel::init()`, the simulation is prepared by reading the input
 files (such as `.yaml`, `.msh`, `particle_locations.csv` files). 
 
 ### DEMModel::integrate()
+
 Key steps in  `DEMModel::integrate()` are 
 ```cpp
 void model::DEMModel::run(inp::Input *deck) {
@@ -197,6 +207,7 @@ void model::DEMModel::integrateVerlet() {
 ```
 
[…]
