+++ CMakeLists.txt
@@ -6,7 +6,7 @@
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
 # file LICENSE)
 
-cmake_minimum_required(VERSION 3.10)
+cmake_minimum_required(VERSION 3.10...4.3)
 
 #----------------------------------#
 #<<<<<<<<<< Project name >>>>>>>>>>#
