### 2025-11-29 prashjha/PeriDEM#45 @prashjha
Hi @matthewfeickert, thank you for the excellent suggestion to use Pixi.

I tried it [here](https://github.com/prashjha/PeriDEM/commit/bab82db1e9c2925eb887afd608dccbe1b47dbb15), and the code works cleanly on Mac. I will verify that this also works on Ubuntu, and then work on the `install` option.
