### 2025-11-29 e0dc3c0
JOSS review: Related to #45, pixi build is verified on mac and ubunut (noble). README includes details of building library using ubuntu docker
