### 2025-11-29 941de34
JOSS review: For #45, the library now supports install.
