+++ src/util/CMakeLists.txt
@@ -11,3 +11,10 @@ AUX_SOURCE_DIRECTORY(./ SOURCES)
 add_library(Util ${SOURCES})
 
 target_link_libraries(Util PUBLIC ${LAPACK_LIBRARIES} ${BLAS_LIBRARIES} fmt::fmt-header-only Threads::Threads MPI::MPI_CXX)
+
+target_include_directories(Util
+  PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src>
+    $<INSTALL_INTERFACE:include>
+)
