+++ src/fe/CMakeLists.txt
@@ -12,3 +12,11 @@ add_library(FE ${SOURCES}
         elemIncludes.h)
 
 target_link_libraries(FE PUBLIC Util RW ${METIS_LIB})
+
+target_include_directories(FE
+  PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src>
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/external>
+    $<INSTALL_INTERFACE:include>
+)
