+++ src/geometry/CMakeLists.txt
@@ -10,4 +10,12 @@ AUX_SOURCE_DIRECTORY(./ SOURCES)
 
 add_library(Geometry ${SOURCES})
 
-target_link_libraries(Geometry PUBLIC Util NSearch fmt::fmt Threads::Threads)
+target_link_libraries(Geometry PUBLIC Util NSearch fmt::fmt-header-only Threads::Threads)
+
+target_include_directories(Geometry
+  PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src>
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/external>
+    $<INSTALL_INTERFACE:include>
+)
