+++ CMakeLists.txt
@@ -11,23 +11,29 @@ cmake_minimum_required(VERSION 3.10...4.3)
 #----------------------------------#
 #<<<<<<<<<< Project name >>>>>>>>>>#
 #----------------------------------#
-project(PeriDEM LANGUAGES CXX C)
+project(PeriDEM VERSION 0.2.1 LANGUAGES CXX C)
 
 set(CMAKE_PROJECT_HOMEPAGE_URL "https://github.com/prashjha/PeriDEM")
 set(CMAKE_PROJECT_DESCRIPTION "PeriDEM -- Peridynamics-based high-fidelity model for granular media")
-
-# version
-set(VERSION_MAJOR 0)
-set(VERSION_MINOR 2)
-set(VERSION_UPDATE 0)
 
 #----------------------------------#
 #<<<<<<<<<< Project setting >>>>>>>>>>#
 #----------------------------------#
 
-# config files
-configure_file("${PROJECT_SOURCE_DIR}/PeriDEMConfig.h.in"
-        "${PROJECT_SOURCE_DIR}/src/PeriDEMConfig.h")
+# Configure header with version info
+configure_file(
+  "${PROJECT_SOURCE_DIR}/PeriDEMConfig.h.in"
+  "${PROJECT_BINARY_DIR}/PeriDEMConfig.h"
+  @ONLY
+)
+
+# Use version from project() for CMake package config
+include(CMakePackageConfigHelpers)
+write_basic_package_version_file(
+  "${CMAKE_CURRENT_BINARY_DIR}/PeriDEMConfigVersion.cmake"
+  VERSION ${PROJECT_VERSION}
+  COMPATIBILITY AnyNewerVersion
+)
 
 # executable directory
 set(EXECUTABLE_OUTPUT_PATH "${PROJECT_BINARY_DIR}/bin")
@@ -159,9 +165,9 @@ endif()
 if (${Enable_CMAKE_Debug_Build})
     message(STATUS "Including key directories")
 endif()
-include_directories(SYSTEM external)
-include_directories(${MPI_INCLUDE_PATH})
-include_directories(src)
+# include_directories(SYSTEM external)
+# include_directories(${MPI_INCLUDE_PATH})
+# include_directories(src)
 
 #----------------------------------#
 #<<<<<<<<<< Subdirectories >>>>>>>>>>#
@@ -212,12 +218,52 @@ endif()
 #----------------------------------#
 add_executable(PeriDEM src/main.cpp)
 
+target_include_directories(PeriDEM
+  PUBLIC
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src>
+    $<BUILD_INTERFACE:${PROJECT_BINARY_DIR}>
+    $<INSTALL_INTERFACE:include>
+)
 target_link_libraries(PeriDEM PUBLIC Model)
+
+install(TARGETS PeriDEM Model Util RW Inp Material Loading Geometry FE Particle NSearch
+  EXPORT PeriDEMTargets
+  RUNTIME DESTINATION bin
+  LIBRARY DESTINATION lib
+  ARCHIVE DESTINATION lib
+)
+
+install(EXPORT PeriDEMTargets
+  FILE PeriDEMTargets.cmake
+  NAMESPACE PeriDEM::
+  DESTINATION lib/cmake/PeriDEM
+)
+
 
 #----------------------------------#
 #<<<<<<<<<< PeriDEM install >>>>>>>>>>#
 #----------------------------------#
-install(TARGETS PeriDEM DESTINATION bin)
+install(FILES ${PROJECT_BINARY_DIR}/PeriDEMConfig.h DESTINATION include)
[…]
