### 2025-11-29 a846f9b
JOSS review: For #45, discuss using installed PeriDEM as package
