### 2025-11-29 4a6fb71
JOSS review: For #45, update README to mention testing install
