### 2025-11-29 9a0e3b7
JOSS review: #45, fix MPI issue in circleci build error
