### 2025-11-29 prashjha/PeriDEM#45 @prashjha
Hi @matthewfeickert, the pull request #48 addresses your comments. Thank you for your thoughtful comments. I learned about Pixi from this, and I think it is very cool.
