+++ namespaceutil.js
@@ -1,149 +1,0 @@
-var namespaceutil =
-[
-    [ "geometry", "namespaceutil_1_1geometry.html", "namespaceutil_1_1geometry" ],
-    [ "io", "namespaceutil_1_1io.html", "namespaceutil_1_1io" ],
-    [ "methods", "namespaceutil_1_1methods.html", [
-      [ "add", "namespaceutil_1_1methods.html#aab17a5d1088344c57c7efd4c06815fe4", null ],
-      [ "addToList", "namespaceutil_1_1methods.html#ad08461b20fe20ffa421967afffd8fbe7", null ],
-      [ "appendKeyData", "namespaceutil_1_1methods.html#a26435993cbfb4aad9e524135d59306d0", null ],
-      [ "getKeyData", "namespaceutil_1_1methods.html#a27516b0cf003c86fffa622abb8e85127", null ],
-      [ "isFree", "namespaceutil_1_1methods.html#ac8ab8ee8690f8baa3e008256f7a5351d", null ],
-      [ "isFree", "namespaceutil_1_1methods.html#a18b94d85f0f6eb12c3f4b9dd462dd325", null ],
-      [ "isInList", "namespaceutil_1_1methods.html#a86460acadab9a2e99a92121c0213e444", null ],
-      [ "isTagInList", "namespaceutil_1_1methods.html#ab21bec324f7d53e91378acc142303fb3", null ],
-      [ "max", "namespaceutil_1_1methods.html#a7ad94e441da65fc0f1cafe183fa252fe", null ],
-      [ "maxAndMaxIndex", "namespaceutil_1_1methods.html#abab275098f631722ec9b85b5277f36a3", null ],
-      [ "maxIndex", "namespaceutil_1_1methods.html#afcbb8d0d56cac1187ad2945bd187f6cf", null ],
-      [ "maxIndex", "namespaceutil_1_1methods.html#ace169ce1b923a51f93e321a512bcf8bf", null ],
-      [ "maxLength", "namespaceutil_1_1methods.html#aff0e69b1eecce494e621eff4f2325f86", null ],
-      [ "maxLengthAndMaxLengthIndex", "namespaceutil_1_1methods.html#a15ca8392b2b1b5f4888d4bda9cdb4473", null ],
-      [ "maxLengthIndex", "namespaceutil_1_1methods.html#a2d8199c377ea884b9471c8b849010625", null ],
-      [ "min", "namespaceutil_1_1methods.html#a801072603311ff11a3cade8cf979a78b", null ],
-      [ "minAndMinIndex", "namespaceutil_1_1methods.html#a36d10e724a3c2d06b047c116ab7ecc2f", null ],
-      [ "minIndex", "namespaceutil_1_1methods.html#a7fd2991e85f113b04b05191b0bf6bfe0", null ],
-      [ "minLength", "namespaceutil_1_1methods.html#a755a4295ae14f1f9adf84fa6eeb3ad29", null ],
-      [ "minLengthAndMinLengthIndex", "namespaceutil_1_1methods.html#aeabd46d5107c0e612523b89614b5764e", null ],
-      [ "minLengthIndex", "namespaceutil_1_1methods.html#a9fb2c7ce477dbb45e4a639d6241c5b53", null ],
-      [ "setKeyData", "namespaceutil_1_1methods.html#af206582cd633d27d42358026631d727a", null ],
-      [ "timeDiff", "namespaceutil_1_1methods.html#aaf9cfac99073d034d41e10d4d17e59a2", null ]
-    ] ],
-    [ "parallel", "namespaceutil_1_1parallel.html", "namespaceutil_1_1parallel" ],
[…]
