+++ namespacerw.js
@@ -1,5 +1,0 @@
-var namespacerw =
-[
-    [ "reader", "namespacerw_1_1reader.html", "namespacerw_1_1reader" ],
-    [ "writer", "namespacerw_1_1writer.html", "namespacerw_1_1writer" ]
-];

[…]
