+++ index.js
@@ -1,0 +1,36 @@
+var index =
+[
+    [ "Table of contents", "index.html#autotoc_md40", null ],
+    [ "Introduction", "index.html#autotoc_md41", null ],
+    [ "Documentation", "index.html#autotoc_md42", null ],
+    [ "Tutorial", "index.html#autotoc_md43", null ],
+    [ "Examples", "index.html#autotoc_md44", [
+      [ "Two-particle tests", "index.html#autotoc_md45", null ],
+      [ "Two-particle with wall test", "index.html#autotoc_md46", null ],
+      [ "Compressive tests", "index.html#autotoc_md47", null ],
+      [ "Attrition tests", "index.html#autotoc_md48", null ],
+      [ "Single particle deformation", "index.html#autotoc_md49", null ]
+    ] ],
+    [ "Brief implementation details", "index.html#autotoc_md50", [
+      [ "DEMModel::run()", "index.html#autotoc_md51", null ],
+      [ "DEMModel::integrate()", "index.html#autotoc_md52", null ],
+      [ "DEMModel::computeForces()", "index.html#autotoc_md53", null ],
+      [ "Further reading", "index.html#autotoc_md54", null ]
+    ] ],
+    [ "Installation", "index.html#autotoc_md55", [
+      [ "Dependencies", "index.html#autotoc_md56", null ],
+      [ "Building the code", "index.html#autotoc_md57", null ],
+      [ "Install & use as a CMake package", "index.html#autotoc_md58", null ],
+      [ "Using docker to test the library", "index.html#autotoc_md59", null ],
+      [ "Future plans", "index.html#autotoc_md60", null ],
+      [ "Ask for help", "index.html#autotoc_md61", null ]
+    ] ],
+    [ "Running simulations", "index.html#autotoc_md62", [
+      [ "Two-particle with wall", "index.html#autotoc_md63", null ],
+      [ "Compressive test", "index.html#autotoc_md64", null ]
+    ] ],
+    [ "Visualizing results", "index.html#autotoc_md65", null ],
+    [ "Contributing", "index.html#autotoc_md66", null ],
+    [ "Citations", "index.html#autotoc_md67", null ],
+    [ "Developers", "index.html#autotoc_md68", null ]
+];

[…]
