### 2025-12-10 2bdf9dc
MNT: Update cmake dependency to allow for CMake policy v4
