### 2025-12-10 dacd8e4
ENH: Add 'start' and additional tasks to Pixi manifest
* Separate tasks out in Pixi manifest to allow for running build, testing,
  and install tasks separately. Add canonical 'start' task to build and
  install software. Add 'test' task that will do a clean build and run
  ctests.
* Update task name from 'build' to 'test' in README.
* Remove 'rm' of build directory from 'build' task to allow for potentially
  skipping files that do not need to be regenerated and to give user slightly
  more control.
