+++ pixi.toml
@@ -4,11 +4,9 @@ name = "PeriDEM"
 platforms = ["linux-64", "osx-arm64"]
 version = "v0.2.1"
 
-# Currently can't install, only build?
-# c.f. https://github.com/prashjha/PeriDEM/blob/8f73cbd12fbf8d483bb46670661f9f2641d03c37/README.md?plain=1#L264
 [tasks.build]
+description = "Build PeriDEM from local source"
 cmd = """
-rm -rf build && \
 cmake   -DEnable_Documentation=OFF \
         -DEnable_Tests=ON \
         -DEnable_High_Load_Tests=OFF \
@@ -20,9 +18,24 @@ cmake   -DEnable_Documentation=OFF \
         -S . \
         -B build && \
 cmake build -LH && \
-cmake --build build --clean-first --parallel "$(nproc --ignore=2)" && \
-ctest --verbose --test-dir build/
+cmake --build build --clean-first --parallel "$(nproc --ignore=2)"
 """
+
+[tasks.ctest]
+description = "Run tests with ctest"
+cmd = "ctest --verbose --test-dir ./build/"
+
+[tasks.install-peridem]
+description = "Install PeriDEM under CMAKE_INSTALL_PREFIX defined in build"
+cmd = "cmake --install build"
+
+[tasks.start]
+description = "Build and install PeriDEM"
+depends-on = ["build", "install-peridem"]
+
+[tasks.test]
+description = "Do a clean build of PeriDEM and run ctest tests"
+depends-on = ["build", "ctest"]
 
 [dependencies]
 cxx-compiler = ">=1.11.0,<2"
