+++ search/pages_0.js
@@ -1,5 +1,0 @@
-var searchData=
-[
-  ['complex_20shaped_20particles_0',['PeriDEM - High-fidelity modeling of granular media consisting of deformable complex-shaped particles',['../index.html',1,'']]],
-  ['consisting_20of_20deformable_20complex_20shaped_20particles_1',['PeriDEM - High-fidelity modeling of granular media consisting of deformable complex-shaped particles',['../index.html',1,'']]]
-];
