+++ examples/PeriDEM/two_particles/twop_concave_and_hex/inp/problem_setup.py
@@ -1,660 +1,0 @@
-import os
-import numpy as np
-import sys
-
-def print_bool(arg, prefix = ""):
-
-  str = prefix
-  if arg == True:
-    str += "True\n"
-  else:
-    str += "False\n"
-  return str
-
-def print_dbl(arg, prefix = ""):
-
-  str = prefix + "%4.6e\n" % (arg)
-  return str
-
-def print_int(arg, prefix = ""):
-  str = prefix + "%d\n" % (arg)
-  return str
-
-def print_dbl_list(arg, prefix = ""):
-  str = prefix + "["
-  N = len(arg)
-  for i in range(N):
-    str += "%4.6e" % (arg[i])
-    if i < N - 1:
-      str += ", "
-    else:
-      str += "]\n"
-
-  return str
-
-def print_int_list(arg, prefix = ""):
-  str = prefix + "["
-  N = len(arg)
-  for i in range(N):
-    str += "%d" % (arg[i])
-    if i < N - 1:
-      str += ", "
-    else:
-      str += "]\n"
-
-  return str
-
-def does_intersect(p, r, R, particles, padding):
-
-  for q in particles:
-
-    pq = np.array([p[i] - q[i] for i in range(3)])
-    if np.linalg.norm(pq) <= r + R + padding:
-      return True
-
-  return False
-
-
-def get_E(K, nu):
-  return 3. * K * (1. - 2. * nu)
-
-def get_G(E, nu):
-  return E / (2. * (1. + nu))
-
-
-def get_eff_k(k1, k2):
-  return 2. * k1 * k2 / (k1 + k2)
-
-
-def particle_locations(inp_dir, pp_tag, R1, R2, offset):
-  """Generate particle location data"""
-
-  sim_particles = []
-  sim_particles.append([0., R1, R1, 0., R1])
-  sim_particles.append([1., R1, 2. * R1 + R2 + offset, 0., R2])
-
-  inpf = open(inp_dir + 'particle_locations_' + str(pp_tag) + '.csv','w')
-  inpf.write("i, x, y, z, r\n")
-  for p in sim_particles:
-    inpf.write("%d, %Lf, %Lf, %Lf, %Lf\n" % (int(p[0]), p[1], p[2], p[3], p[4]))
-
-  inpf.close()
-
-def particle_locations_orient(inp_dir, pp_tag, R1, R2, offset):
-  """Generate particle location data"""
-
-  sim_particles = []
-  sim_particles.append([0., R1, R1, 0., R1, 0.5*np.pi])
-  sim_particles.append([1., R1, 2. * R1 + R2 + offset, 0., R2, 0.])
-
-  inpf = open(inp_dir + 'particle_locations_' + str(pp_tag) + '.csv','w')
-  inpf.write("i, x, y, z, r, o\n")
-  for p in sim_particles:
-    inpf.write("%d, %Lf, %Lf, %Lf, %Lf, %Lf\n" % (int(p[0]), p[1], p[2], p[3], p[4], p[5]))
-
-  inpf.close()
-
-def rotate(p, theta, axis):
-
-  p_np = np.array(p)
-  axis_np = np.array(axis)
-
-  ct = np.cos(theta);
-  st = np.sin(theta);
-
-  # dot
-  p_dot_n = np.dot(p_np,axis_np)
-
-  # cross
-  n_cross_p = np.cross(axis_np, p_np)
-
-  return (1. - ct) * p_dot_n * axis_np + ct * p_np + st * n_cross_p
-
-
-def get_ref_hex_points(center, radius, add_center = False):
-
-  # drum2d
-  #
-  #                        v3           v2
-  #                         +           +
-  #
-  #
[…]
