+++ search/all_7.js
@@ -1,13 +1,0 @@
-var searchData=
-[
-  ['h_0',['h',['../namespacegmsh__particles.html#ad9e6465f12f3fa0dd2db6f3579f2c5f2',1,'gmsh_particles.h'],['../namespaceparticle__wall.html#a928cfa4914ddbebc3278ad7f77e540d8',1,'particle_wall.h']]],
-  ['hatfunction_1',['hatFunction',['../namespaceutil.html#a51adb530fc272317c3a859af27787b1a',1,'util']]],
-  ['hatfunctionquick_2',['hatFunctionQuick',['../namespaceutil.html#a7ad8c95bec3f5405435dff75422e6993',1,'util']]],
-  ['help_3',['help',['../index.html#autotoc_md61',1,'Ask for help'],['../namespaceparticle__wall.html#a1da122babcef1b91f502d7b473b9daef',1,'particle_wall.help']]],
-  ['hexagon_4',['hexagon',['../classutil_1_1geometry_1_1Hexagon.html',1,'util::geometry::Hexagon'],['../classutil_1_1geometry_1_1Hexagon.html#a0b7179b6eac1d4aa4cc9aa10a9cdb5d3',1,'util::geometry::Hexagon::Hexagon()'],['../classutil_1_1geometry_1_1Hexagon.html#a87fcb9a61198f2194e1090757efa8f38',1,'util::geometry::Hexagon::Hexagon(double r, util::Point x=util::Point(0., 0., 0.), util::Point a=util::Point(1., 0., 0.), std::string description=&quot;&quot;)']]],
-  ['hexagon_5fmesh_5fsymmetric_5',['hexagon_mesh_symmetric',['../namespacegmsh__particles.html#a4ad630d57e69747a673e2191b8d68bf9',1,'gmsh_particles']]],
-  ['high_20fidelity_20modeling_20of_20granular_20media_20consisting_20of_20deformable_20complex_20shaped_20particles_6',['PeriDEM - High-fidelity modeling of granular media consisting of deformable complex-shaped particles',['../index.html',1,'']]],
-  ['hole_5flx_7',['hole_Lx',['../namespacegmsh__particles.html#a27edb5fdee61f35bca67ed2b355bb36a',1,'gmsh_particles']]],
-  ['hole_5fly_8',['hole_Ly',['../namespacegmsh__particles.html#a93cc2737139ea9034fe94b0911e208b2',1,'gmsh_particles']]],
-  ['horizon_9',['horizon',['../namespaceparticle__wall.html#aa08c5c90a5cffb178f26216ae2205144',1,'particle_wall']]]
-];
