+++ search/pages_1.js
@@ -1,5 +1,0 @@
-var searchData=
-[
-  ['deformable_20complex_20shaped_20particles_0',['PeriDEM - High-fidelity modeling of granular media consisting of deformable complex-shaped particles',['../index.html',1,'']]],
-  ['dockerimages_1',['dockerimages',['../md__2home_2user_2project_2tools_2docker_2README.html',1,'']]]
-];
