+++ search/files_c.js
@@ -1,6 +1,0 @@
-var searchData=
-[
-  ['quaddata_2eh_0',['quadData.h',['../quadData_8h.html',1,'']]],
-  ['quadelem_2ecpp_1',['quadElem.cpp',['../quadElem_8cpp.html',1,'']]],
-  ['quadelem_2eh_2',['quadElem.h',['../quadElem_8h.html',1,'']]]
-];
