+++ structinp_1_1PBCData.js
@@ -1,17 +1,0 @@
-var structinp_1_1PBCData =
-[
-    [ "PBCData", "structinp_1_1PBCData.html#aec2ba88de2aaa0d3fcb9451df0475529", null ],
-    [ "print", "structinp_1_1PBCData.html#a9ea2acc89b05ecd9fcc7eab46ca944fa", null ],
-    [ "printStr", "structinp_1_1PBCData.html#a22f266c6844df9da79ce450db2edee41", null ],
-    [ "d_direction", "structinp_1_1PBCData.html#adb7cb746bb9847d3965de08f29c962fd", null ],
-    [ "d_isDisplacementZero", "structinp_1_1PBCData.html#ab1f29b184f40702f73f4692d12b08161", null ],
-    [ "d_isRegionActive", "structinp_1_1PBCData.html#abc9699c0f1c959bb7cf2b319a10806d0", null ],
-    [ "d_pList", "structinp_1_1PBCData.html#a59f63d23e705e9dd72489487fd855ce0", null ],
-    [ "d_pNotList", "structinp_1_1PBCData.html#aae4c4f0b9af863a27ee0436032a6afbe", null ],
-    [ "d_regionGeomData", "structinp_1_1PBCData.html#a8599685e20aab035eee9a917e3e8dd9b", null ],
-    [ "d_selectionType", "structinp_1_1PBCData.html#a7d2951bae622e6be9348c450559403ad", null ],
-    [ "d_spatialFnParams", "structinp_1_1PBCData.html#a672fb591f177bad9109141270e7c7a78", null ],
-    [ "d_spatialFnType", "structinp_1_1PBCData.html#a8f5790a81092f814e9083068ef9352b5", null ],
-    [ "d_timeFnParams", "structinp_1_1PBCData.html#abd134455f1c6e1a580c918a35314e984", null ],
-    [ "d_timeFnType", "structinp_1_1PBCData.html#a1ceb76a1c62e9067163692c9d67805fb", null ]
-];
