+++ search/typedefs_0.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['coord_5ft_0',['coord_t',['../structnsearch_1_1PointCloudAdaptor.html#acc4441c56d7925bae17969f9a44abf82',1,'nsearch::PointCloudAdaptor']]]
-];

[…]
