### 2025-12-10 c04581b
automated deployment to GitHub Pages: 58737c3313f887b8657607ddb764033bdd1c3cf0
