### 2025-12-12 prashjha/PeriDEM#50 @matthewfeickert
[JOSS Review]: Is "Brief Introduction to PeriDEM Model" section necessary?
Following https://github.com/openjournals/joss-reviews/issues/7525#issuecomment-3642764651, I've read the current state of the paper draft in [`docs/joss/paper.md` on the `joss` branch](https://github.com/prashjha/PeriDEM/blob/5d3d2f9ef05f8d303f4b2f84e5881056b2b93703/docs/joss/paper.md). I think things are good, though I'm not certain that the section ["Brief Introduction to PeriDEM Model"](https://github.com/prashjha/PeriDEM/blob/5d3d2f9ef05f8d303f4b2f84e5881056b2b93703/docs/joss/paper.md#L38-L64) is needed, given that this is explaining the physics model itself and the [JOSS docs](https://joss.readthedocs.io/en/latest/paper.html) [note](https://github.com/openjournals/joss/blob/8b124befb9d2ce0eba69857be3029a65c0dc721f/docs/paper.md?plain=1#L15-L22)

> Your paper should include:
>
> - A list of the authors of the software and their affiliations, using the correct format (see the example below).
> - A summary describing the high-level functionality and purpose of the software for a diverse, *non-specialist audience*.
> - A *Statement of need* section that clearly illustrates the research purpose of the software and places it in the context of related work.
> - A list of key references, including to other software addressing related needs. Note that the references should include full names of venues, e.g., journals and conferences, not abbreviations only understood in the context of a specific discipline.
> - Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.
> - Acknowledgement of any financial support.

and this section feels both more technical and isn't listed as a requirement.

My personal thoughts are that this section should be moved _to the documentation_. As far as I can tell this level of detail about the physics model doesn't exist in the docs yet and it only present in [doi.org/10.1016/j.jmps.2021.104376](https://doi.org/10.1016/j.jmps.2021.104376).

I also don't think we're even able to use Figure 2 in the paper

https://github.com/prashjha/PeriDEM/blob/5d3d2f9ef05f8d303f4b2f84e5881056b2b93703/docs/joss/paper.md?plain=1#L64

without permission from Elsevier, as those are images used in Figure 2 of [doi.org/10.1016/j.jmps.2021.104376](https://doi.org/10.1016/j.jmps.2021.104376) and so Elsevier now owns the copyright to them (unless you self published them in a Zenodo archive or something before publication under a cc license and then cited the archive when using them). Note that I would be very happy to be wrong here. This is a conversation I've had with @kyleniemeyer a few times, and he might have a more informed view on it than I do.

What are you general thoughts on the "Brief Introduction to PeriDEM Model" section? You wrote it and so I don't want to be dismissive of that information, and I think it is nice to have associated with the software, just maybe not in the JOSS paper.
