### 2025-12-12 prashjha/PeriDEM#50 @prashjha
Hi @matthewfeickert, I would prefer to keep the 'Brief Introduction to PeriDEM Model' in the JOSS paper, and I also liked the idea of including such descriptions in the library documentation.  The attached figure differs slightly from the one in the paper. I have used this figure in many of my conferences. I don't see any copyright issues with this figure.
