### 2025-12-16 prashjha/PeriDEM#50 @matthewfeickert
> I would prefer to keep the 'Brief Introduction to PeriDEM Model' in the JOSS paper, and I also liked the idea of including such descriptions in the library documentation.

Okay, that section can stay but I don't think that

> The attached figure differs slightly from the one in the paper.

matters in terms of copyright.

> I have used this figure in many of my conferences. I don't see any copyright issues with this figure.

When you published https://doi.org/10.1016/j.jmps.2021.104376 you gave full copyright of the contents and any derived work to Elsevier, as the publication is not open access. The Elsevier Permissions page (https://www.elsevier.com/about/policies-and-standards/copyright/permissions) includes publication of published images explicitly in this

> As a general rule, permission should be sought from the rights holder to reproduce any substantial part of a copyrighted work. This includes any text, illustrations, charts, tables, photographs, or other material from previously published sources. Obtaining permission to re-use content published by Elsevier is simple. Follow the guide below for a quick and easy route to permission.

> Permission is required for any material that is not original. As a rule, written permission must be obtained from the rightsholder to re-use any copyrighted material. Typically, the rightsholder of published material is the publisher unless it is explicitly indicated otherwise. Copyrighted material can include figures, illustrations, charts, tables, photographs, and text excerpts. Re-use of any borrowed material must be properly acknowledged, even if it is determined that written permission is not necessary.

As is, the JOSS draft reuses elements from or makes small edits to the published Elsevier work for

| [JOSS Draft](https://github.com/prashjha/PeriDEM/tree/5d3d2f9ef05f8d303f4b2f84e5881056b2b93703) Figure | https://doi.org/10.1016/j.jmps.2021.104376 Figure |
| --- | --- |
| 2 | 2 |
| 3a | 22 |
| 3b | 24 |

(Also, while https://doi.org/10.1016/j.jmps.2021.104376 is cited in the JOSS draft, the figure text makes no reference to the publication.)

Other members of the JOSS editorial team agree that without explicit permissions from the publisher these figures can not be published under JOSS. Though as we're of a difference of opinion on this I would fully encourage you to also check with the EiC team. As we're very close to publishing this JOSS submission it would be great if this could get resolved this week so that the submission can be published before next week.

**edit**: I see now that you did publish a preprint on arXiv (https://arxiv.org/abs/2010.07218) 👍 , but it uses the [arXiv.org perpetual, non-exclusive license 1.0](https://arxiv.org/licenses/nonexclusive-distrib/1.0/), which is not a CC BY-SA license.
