### 2025-12-17 20fe4b3
Update figures with adaptation notes in paper.md #50
