### 2025-12-17 c8c09d9
fix reference inside figure caption #50
