### 2025-12-17 e0c5b6d
Update figure caption formatting in paper.md #50
