### 2025-12-17 prashjha/PeriDEM#50 @prashjha
Hi @matthewfeickert, I contacted Elsevier (see some more details in the Slack channel), and since I am the author of the original article, I retain the rights to reuse and adapt my work as done in this article. They did ask me to acknowledge the published article properly. I already acknowledged the article, but just to be safe, I updated the figure captions to include a reference. I hope this helps address the issue. Once you let me know, I will close it. Thank you for being very diligent about this!
