### 2025-12-18 prashjha/PeriDEM#50 @matthewfeickert
Resolved by https://github.com/prashjha/PeriDEM/commit/20fe4b391bca694389b3f9276ded198f96095854.

Thanks for all the work you did here in short order, @prashjha. You went above and beyond also sharing everything you learned with the rest of the editorial team, which is going to be quite useful in the future I think. 👍
